Struct rune::alloc::prelude::Vec

pub struct Vec<T, A = Global>
where
    A: Allocator,
{ /* private fields */ }

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Examples

use rune::alloc::Vec;
use rune::alloc::prelude::*;

let mut vec = Vec::new();
vec.try_push(1)?;
vec.try_push(2)?;

assert_eq!(vec.len(), 2);
assert_eq!(vec[0], 1);

assert_eq!(vec.pop(), Some(2));
assert_eq!(vec.len(), 1);

vec[0] = 7;
assert_eq!(vec[0], 7);

vec.try_extend([1, 2, 3])?;

for x in &vec {
    println!("{x}");
}

assert_eq!(vec, [7, 1, 2, 3]);

The try_vec! macro is provided for convenient initialization:

use rune::alloc::{try_vec, Vec};

let mut vec1 = try_vec![1, 2, 3];
vec1.try_push(4)?;
let vec2 = Vec::try_from([1, 2, 3, 4])?;
assert_eq!(vec1, vec2);

It can also initialize each element of a Vec<T> with a given value. This may be more efficient than performing allocation and initialization in separate steps, especially when initializing a vector of zeros:

use rune::alloc::{try_vec, Vec};

let vec = try_vec![0; 5];
assert_eq!(vec, [0, 0, 0, 0, 0]);

// The following is equivalent, but potentially slower:
let mut vec = Vec::try_with_capacity(5)?;
vec.try_resize(5, 0)?;
assert_eq!(vec, [0, 0, 0, 0, 0]);

For more information, see Capacity and Reallocation.

Use a Vec<T> as an efficient stack:

use rune::alloc::Vec;

let mut stack = Vec::new();

stack.try_push(1)?;
stack.try_push(2)?;
stack.try_push(3)?;

while let Some(top) = stack.pop() {
    // Prints 3, 2, 1
    println!("{top}");
}

Indexing

The Vec type allows to access values by index, because it implements the Index trait. An example will be more explicit:

use rune::alloc::try_vec;

let v = try_vec![0, 2, 4, 6];
println!("{}", v[1]); // it will display '2'

However be careful: if you try to access an index which isn’t in the Vec, your software will panic! You cannot do this:

use rune::alloc::try_vec;

let v = try_vec![0, 2, 4, 6];
println!("{}", v[6]); // it will panic!

Use get and get_mut if you want to check whether the index is in the Vec.

Slicing

A Vec can be mutable. On the other hand, slices are read-only objects. To get a slice, use &. Example:

use rune::alloc::try_vec;

fn read_slice(slice: &[usize]) {
    // ...
}

let v = try_vec![0, 1];
read_slice(&v);

// ... and that's all!
// you can also do it like this:
let u: &[usize] = &v;
// or like this:
let u: &[_] = &v;

In Rust, it’s more common to pass slices as arguments rather than vectors when you just want to provide read access. The same goes for String and &str.

Capacity and reallocation

The capacity of a vector is the amount of space allocated for any future elements that will be added onto the vector. This is not to be confused with the length of a vector, which specifies the number of actual elements within the vector. If a vector’s length exceeds its capacity, its capacity will automatically be increased, but its elements will have to be reallocated.

For example, a vector with capacity 10 and length 0 would be an empty vector with space for 10 more elements. Pushing 10 or fewer elements onto the vector will not change its capacity or cause reallocation to occur. However, if the vector’s length is increased to 11, it will have to reallocate, which can be slow. For this reason, it is recommended to use Vec::try_with_capacity whenever possible to specify how big the vector is expected to get.

Guarantees

Due to its incredibly fundamental nature, Vec makes a lot of guarantees about its design. This ensures that it’s as low-overhead as possible in the general case, and can be correctly manipulated in primitive ways by unsafe code. Note that these guarantees refer to an unqualified Vec<T>. If additional type parameters are added (e.g., to support custom allocators), overriding their defaults may change the behavior.

Most fundamentally, Vec is and always will be a (pointer, capacity, length) triplet. No more, no less. The order of these fields is completely unspecified, and you should use the appropriate methods to modify these. The pointer will never be null, so this type is null-pointer-optimized.

However, the pointer might not actually point to allocated memory. In particular, if you construct a Vec with capacity 0 via Vec::new, [try_vec![]], [Vec::try_with_capacity(0)], or by calling try_shrink_to_fit on an empty Vec, it will not allocate memory. Similarly, if you store zero-sized types inside a Vec, it will not allocate space for them. Note that in this case the Vec might not report a capacity of 0. Vec will allocate if and only if mem::size_of::<T>() * capacity() > 0. In general, Vec’s allocation details are very subtle — if you intend to allocate memory using a Vec and use it for something else (either to pass to unsafe code, or to build your own memory-backed collection), be sure to deallocate this memory by using from_raw_parts to recover the Vec and then dropping it.

[try_vec![]]: try_vec! [Vec::try_with_capacity(0)]: Vec::try_with_capacity

If a Vec has allocated memory, then the memory it points to is on the heap (as defined by the allocator Rust is configured to use by default), and its pointer points to len initialized, contiguous elements in order (what you would see if you coerced it to a slice), followed by capacity - len logically uninitialized, contiguous elements.

A vector containing the elements 'a' and 'b' with capacity 4 can be visualized as below. The top part is the Vec struct, it contains a pointer to the head of the allocation in the heap, length and capacity. The bottom part is the allocation on the heap, a contiguous memory block.

            ptr      len  capacity
       +--------+--------+--------+
       | 0x0123 |      2 |      4 |
       +--------+--------+--------+
            |
            v
Heap   +--------+--------+--------+--------+
       |    'a' |    'b' | uninit | uninit |
       +--------+--------+--------+--------+

uninit represents memory that is not initialized, see MaybeUninit.

Note: the ABI is not stable and Vec makes no guarantees about its memory layout (including the order of fields).

Vec will never perform a “small optimization” where elements are actually stored on the stack for two reasons:

It would make it more difficult for unsafe code to correctly manipulate a Vec. The contents of a Vec wouldn’t have a stable address if it were only moved, and it would be more difficult to determine if a Vec had actually allocated memory.

It would penalize the general case, incurring an additional branch on every access.

Vec will never automatically shrink itself, even if completely empty. This ensures no unnecessary allocations or deallocations occur. Emptying a Vec and then filling it back up to the same len should incur no calls to the allocator. If you wish to free up unused memory, use try_shrink_to_fit or try_shrink_to.

try_push and try_insert will never (re)allocate if the reported capacity is sufficient. try_push and try_insert will (re)allocate if len == capacity. That is, the reported capacity is completely accurate, and can be relied on. It can even be used to manually free the memory allocated by a Vec if desired. Bulk insertion methods may reallocate, even when not necessary.

Vec does not guarantee any particular growth strategy when reallocating when full, nor when try_reserve is called. The current strategy is basic and it may prove desirable to use a non-constant growth factor. Whatever strategy is used will of course guarantee O(1) amortized try_push.

try_vec![x; n], try_vec![a, b, c, d], and Vec::try_with_capacity(n), will all produce a Vec with exactly the requested capacity. If len == capacity, (as is the case for the try_vec! macro), then a Vec<T> can be converted to and from a Box<[T]> without reallocating or moving the elements.

Vec will not specifically overwrite any data that is removed from it, but also won’t specifically preserve it. Its uninitialized memory is scratch space that it may use however it wants. It will generally just do whatever is most efficient or otherwise easy to implement. Do not rely on removed data to be erased for security purposes. Even if you drop a Vec, its buffer may simply be reused by another allocation. Even if you zero a Vec’s memory first, that might not actually happen because the optimizer does not consider this a side-effect that must be preserved. There is one case which we will not break, however: using unsafe code to write to the excess capacity, and then increasing the length to match, is always valid.

Currently, Vec does not guarantee the order in which elements are dropped. The order has changed in the past and may change again.

Implementations

impl<T> Vec<T>

pub const fn new() -> Vec<T>

Constructs a new, empty Vec<T>.

The vector will not allocate until elements are pushed onto it.

Examples

let mut vec: Vec<i32> = Vec::new();

pub fn try_with_capacity(capacity: usize) -> Result<Vec<T>, Error>

Constructs a new, empty Vec<T> with at least the specified capacity.

The vector will be able to hold at least capacity elements without reallocating. This method is allowed to allocate for more elements than capacity. If capacity is 0, the vector will not allocate.

It is important to note that although the returned vector has the minimum capacity specified, the vector will have a zero length. For an explanation of the difference between length and capacity, see Capacity and reallocation.

If it is important to know the exact allocated capacity of a Vec, always use the capacity method after construction.

For Vec<T> where T is a zero-sized type, there will be no allocation and the capacity will always be usize::MAX.

Panics

Panics if the new capacity exceeds isize::MAX bytes.

Examples

let mut vec = Vec::with_capacity(10);

// The vector contains no items, even though it has capacity for more
assert_eq!(vec.len(), 0);
assert!(vec.capacity() >= 10);

// These are all done without reallocating...
for i in 0..10 {
    vec.push(i);
}
assert_eq!(vec.len(), 10);
assert!(vec.capacity() >= 10);

// ...but this may make the vector reallocate
vec.push(11);
assert_eq!(vec.len(), 11);
assert!(vec.capacity() >= 11);

// A vector of a zero-sized type will always over-allocate, since no
// allocation is necessary
let vec_units = Vec::<()>::with_capacity(10);
assert_eq!(vec_units.capacity(), usize::MAX);

pub fn into_std(self) -> Vec<T>

Convert a Vec<T> into a std Vec<T>.

The result is allocated on the heap, using the default global allocator so this is a zero-copy operation.

The memory previously occupied by this vector will be released.

impl<T, A> Vec<T, A>
where A: Allocator,

pub const fn new_in(alloc: A) -> Vec<T, A>

Constructs a new, empty Vec<T, A>.

The vector will not allocate until elements are pushed onto it.

Examples

use rune::alloc::Vec;
use rune::alloc::alloc::Global;

let mut vec: Vec<i32, Global> = Vec::new_in(Global);

pub fn try_with_capacity_in( capacity: usize, alloc: A, ) -> Result<Vec<T, A>, Error>

Constructs a new, empty Vec<T, A> with at least the specified capacity with the provided allocator.

The vector will be able to hold at least capacity elements without reallocating. This method is allowed to allocate for more elements than capacity. If capacity is 0, the vector will not allocate.

It is important to note that although the returned vector has the minimum capacity specified, the vector will have a zero length. For an explanation of the difference between length and capacity, see Capacity and reallocation.

If it is important to know the exact allocated capacity of a Vec, always use the capacity method after construction.

For Vec<T, A> where T is a zero-sized type, there will be no allocation and the capacity will always be usize::MAX.

Errors

Errors with Error::CapacityOverflow if the new capacity exceeds isize::MAX bytes.

Examples

use rune::alloc::Vec;
use rune::alloc::alloc::Global;

let mut vec = Vec::try_with_capacity_in(10, Global)?;

// The vector contains no items, even though it has capacity for more
assert_eq!(vec.len(), 0);
assert!(vec.capacity() >= 10);

// These are all done without reallocating...
for i in 0..10 {
    vec.try_push(i)?;
}

assert_eq!(vec.len(), 10);
assert!(vec.capacity() >= 10);

// ...but this may make the vector reallocate
vec.try_push(11)?;
assert_eq!(vec.len(), 11);
assert!(vec.capacity() >= 11);

// A vector of a zero-sized type will always over-allocate, since no
// allocation is necessary
let vec_units = Vec::<(), Global>::try_with_capacity_in(10, Global)?;
assert_eq!(vec_units.capacity(), usize::MAX);

pub unsafe fn from_raw_parts_in( ptr: *mut T, length: usize, capacity: usize, alloc: A, ) -> Vec<T, A>

Creates a Vec<T, A> directly from a pointer, a capacity, a length, and an allocator.

Safety

This is highly unsafe, due to the number of invariants that aren’t checked:

ptr must be currently allocated via the given allocator alloc.

T needs to have the same alignment as what ptr was allocated with. (T having a less strict alignment is not sufficient, the alignment really needs to be equal to satisfy the dealloc requirement that memory must be allocated and deallocated with the same layout.)

The size of T times the capacity (ie. the allocated size in bytes) needs to be the same size as the pointer was allocated with. (Because similar to alignment, dealloc must be called with the same layout size.)

length needs to be less than or equal to capacity.

The first length values must be properly initialized values of type T.

capacity needs to fit the layout size that the pointer was allocated with.

The allocated size in bytes must be no larger than isize::MAX. See the safety documentation of pointer::offset.

These requirements are always upheld by any ptr that has been allocated via Vec<T, A>. Other allocation sources are allowed if the invariants are upheld.

Violating these may cause problems like corrupting the allocator’s internal data structures. For example it is not safe to build a Vec<u8> from a pointer to a C char array with length size_t. It’s also not safe to build one from a Vec<u16> and its length, because the allocator cares about the alignment, and these two types have different alignments. The buffer was allocated with alignment 2 (for u16), but after turning it into a Vec<u8> it’ll be deallocated with alignment 1.

The ownership of ptr is effectively transferred to the Vec<T> which may then deallocate, reallocate or change the contents of memory pointed to by the pointer at will. Ensure that nothing else uses the pointer after calling this function.

Examples

use std::ptr;
use std::mem;

use rune::alloc::Vec;
use rune::alloc::alloc::Global;

let mut v = Vec::try_with_capacity_in(3, Global)?;
v.try_push(1)?;
v.try_push(2)?;
v.try_push(3)?;

// Prevent running `v`'s destructor so we are in complete control
// of the allocation.
let mut v = mem::ManuallyDrop::new(v);

// Pull out the various important pieces of information about `v`
let p = v.as_mut_ptr();
let len = v.len();
let cap = v.capacity();
let alloc = v.allocator();

unsafe {
    // Overwrite memory with 4, 5, 6
    for i in 0..len {
        ptr::write(p.add(i), 4 + i);
    }

    // Put everything back together into a Vec
    let rebuilt = Vec::from_raw_parts_in(p, len, cap, alloc.clone());
    assert_eq!(rebuilt, [4, 5, 6]);
}

Using memory that was allocated elsewhere:

use core::alloc::Layout;

use rune::alloc::Vec;
use rune::alloc::alloc::{Allocator, AllocError, Global};

let layout = Layout::array::<u32>(16).expect("overflow cannot happen");

let vec = unsafe {
    let mem = match Global.allocate(layout) {
        Ok(mem) => mem.cast::<u32>().as_ptr(),
        Err(AllocError) => return,
    };

    mem.write(1_000_000);

    Vec::from_raw_parts_in(mem, 1, 16, Global)
};

assert_eq!(vec, &[1_000_000]);
assert_eq!(vec.capacity(), 16);

pub fn allocator(&self) -> &A

Returns a reference to the underlying allocator.

pub fn into_raw_parts_with_alloc(self) -> (*mut T, usize, usize, A)

Decomposes a Vec<T> into its raw components.

Returns the raw pointer to the underlying data, the length of the vector (in elements), the allocated capacity of the data (in elements), and the allocator. These are the same arguments in the same order as the arguments to from_raw_parts_in.

After calling this function, the caller is responsible for the memory previously managed by the Vec. The only way to do this is to convert the raw pointer, length, and capacity back into a Vec with the from_raw_parts_in function, allowing the destructor to perform the cleanup.

Examples

use rune::alloc::Vec;
use rune::alloc::alloc::Global;

let mut v: Vec<i32> = Vec::new_in(Global);
v.try_push(-1)?;
v.try_push(0)?;
v.try_push(1)?;

let (ptr, len, cap, alloc) = v.into_raw_parts_with_alloc();

let rebuilt = unsafe {
    // We can now make changes to the components, such as
    // transmuting the raw pointer to a compatible type.
    let ptr = ptr as *mut u32;

    Vec::from_raw_parts_in(ptr, len, cap, alloc)
};

assert_eq!(rebuilt, [4294967295, 0, 1]);

pub fn capacity(&self) -> usize

Returns the total number of elements the vector can hold without reallocating.

Examples

use rune::alloc::Vec;
use rune::alloc::alloc::Global;

let mut vec: Vec<i32> = Vec::try_with_capacity_in(10, Global)?;
vec.try_push(42)?;
assert!(vec.capacity() >= 10);

pub fn try_reserve(&mut self, additional: usize) -> Result<(), Error>

Tries to reserve capacity for at least additional more elements to be inserted in the given Vec<T>. The collection may reserve more space to speculatively avoid frequent reallocations. After calling try_reserve, capacity will be greater than or equal to self.len() + additional if it returns Ok(()). Does nothing if capacity is already sufficient. This method preserves the contents even if an error occurs.

Errors

If the capacity overflows, or the allocator reports a failure, then an error is returned.

Examples

use rune::alloc::{Vec, Error};

fn process_data(data: &[u32]) -> Result<Vec<u32>, Error> {
    let mut output = Vec::new();

    // Pre-reserve the memory, exiting if we can't
    output.try_reserve(data.len())?;

    for value in data {
       output.try_push(*value)?;
    }

    Ok(output)
}

pub fn try_reserve_exact(&mut self, additional: usize) -> Result<(), Error>

Tries to reserve the minimum capacity for at least additional elements to be inserted in the given Vec<T>. Unlike try_reserve, this will not deliberately over-allocate to speculatively avoid frequent allocations. After calling try_reserve_exact, capacity will be greater than or equal to self.len() + additional if it returns Ok(()). Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore, capacity can not be relied upon to be precisely minimal. Prefer try_reserve if future insertions are expected.

Errors

If the capacity overflows, or the allocator reports a failure, then an error is returned.

Examples

use rune::alloc::{Vec, Error};
use rune::alloc::prelude::*;

fn process_data(data: &[u32]) -> Result<Vec<u32>, Error> {
    let mut output = Vec::new();

    // Pre-reserve the memory, exiting if we can't
    output.try_reserve_exact(data.len())?;

    // Now we know this can't OOM in the middle of our complex work
    output.try_extend(data.iter().map(|&val| {
        val * 2 + 5 // very complicated
    }))?;

    Ok(output)
}

pub fn try_shrink_to_fit(&mut self) -> Result<(), Error>

Shrinks the capacity of the vector as much as possible.

It will drop down as close as possible to the length but the allocator may still inform the vector that there is space for a few more elements.

Examples

use rune::alloc::Vec;
use rune::alloc::prelude::*;

let mut vec = Vec::try_with_capacity(10)?;
vec.try_extend([1, 2, 3])?;
assert!(vec.capacity() >= 10);
vec.try_shrink_to_fit()?;
assert!(vec.capacity() >= 3);

pub fn try_shrink_to(&mut self, min_capacity: usize) -> Result<(), Error>

Shrinks the capacity of the vector with a lower bound.

The capacity will remain at least as large as both the length and the supplied value.

If the current capacity is less than the lower limit, this is a no-op.

Examples

use rune::alloc::Vec;
use rune::alloc::prelude::*;

let mut vec = Vec::try_with_capacity(10)?;
vec.try_extend([1, 2, 3])?;
assert!(vec.capacity() >= 10);
vec.try_shrink_to(4)?;
assert!(vec.capacity() >= 4);
vec.try_shrink_to(0)?;
assert!(vec.capacity() >= 3);

pub fn try_into_boxed_slice(self) -> Result<Box<[T], A>, Error>

Converts the vector into Box<[T]>.

If the vector has excess capacity, its items will be moved into a newly-allocated buffer with exactly the right capacity.

Examples

use rune::alloc::try_vec;

let v = try_vec![1, 2, 3];
let slice = v.try_into_boxed_slice()?;

Any excess capacity is removed:

use rune::alloc::Vec;
use rune::alloc::prelude::*;

let mut vec = Vec::try_with_capacity(10)?;
vec.try_extend([1, 2, 3])?;

assert!(vec.capacity() >= 10);
let slice = vec.try_into_boxed_slice()?;
assert_eq!(Vec::from(slice).capacity(), 3);

pub fn truncate(&mut self, len: usize)

Shortens the vector, keeping the first len elements and dropping the rest.

If len is greater than the vector’s current length, this has no effect.

The drain method can emulate truncate, but causes the excess elements to be returned instead of dropped.

Note that this method has no effect on the allocated capacity of the vector.

Examples

Truncating a five element vector to two elements:

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 3, 4, 5];
vec.truncate(2);
assert_eq!(vec, [1, 2]);

No truncation occurs when len is greater than the vector’s current length:

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 3];
vec.truncate(8);
assert_eq!(vec, [1, 2, 3]);

Truncating when len == 0 is equivalent to calling the clear method.

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 3];
vec.truncate(0);
assert_eq!(vec, []);

pub fn as_slice(&self) -> &[T]

Extracts a slice containing the entire vector.

Equivalent to &s[..].

Examples

use std::io::{self, Write};
use rune::alloc::try_vec;

let buffer = try_vec![1, 2, 3, 5, 8];
io::sink().write(buffer.as_slice()).unwrap();

pub fn as_mut_slice(&mut self) -> &mut [T]

Extracts a mutable slice of the entire vector.

Equivalent to &mut s[..].

Examples

use std::io::{self, Read};
use rune::alloc::try_vec;

let mut buffer = try_vec![0; 3];
io::repeat(0b101).read_exact(buffer.as_mut_slice()).unwrap();

pub fn as_ptr(&self) -> *const T

Returns a raw pointer to the vector’s buffer, or a dangling raw pointer valid for zero sized reads if the vector didn’t allocate.

The caller must ensure that the vector outlives the pointer this function returns, or else it will end up pointing to garbage. Modifying the vector may cause its buffer to be reallocated, which would also make any pointers to it invalid.

The caller must also ensure that the memory the pointer (non-transitively) points to is never written to (except inside an UnsafeCell) using this pointer or any pointer derived from it. If you need to mutate the contents of the slice, use as_mut_ptr.

Examples

use rune::alloc::try_vec;

let x = try_vec![1, 2, 4];
let x_ptr = x.as_ptr();

unsafe {
    for i in 0..x.len() {
        assert_eq!(*x_ptr.add(i), 1 << i);
    }
}

pub fn as_mut_ptr(&mut self) -> *mut T

Returns an unsafe mutable pointer to the vector’s buffer, or a dangling raw pointer valid for zero sized reads if the vector didn’t allocate.

The caller must ensure that the vector outlives the pointer this function returns, or else it will end up pointing to garbage. Modifying the vector may cause its buffer to be reallocated, which would also make any pointers to it invalid.

Examples

use rune::alloc::Vec;

// Allocate vector big enough for 4 elements.
let size = 4;
let mut x: Vec<i32> = Vec::try_with_capacity(size)?;
let x_ptr = x.as_mut_ptr();

// Initialize elements via raw pointer writes, then set length.
unsafe {
    for i in 0..size {
        *x_ptr.add(i) = i as i32;
    }
    x.set_len(size);
}
assert_eq!(&*x, &[0, 1, 2, 3]);

pub unsafe fn set_len(&mut self, new_len: usize)

Forces the length of the vector to new_len.

This is a low-level operation that maintains none of the normal invariants of the type. Normally changing the length of a vector is done using one of the safe operations instead, such as truncate, try_resize, try_extend, or clear.

Safety

new_len must be less than or equal to capacity().

The elements at old_len..new_len must be initialized.

Examples

This method can be useful for situations in which the vector is serving as a buffer for other code, particularly over FFI:

pub(crate) fn get_dictionary(&self) -> Option<Vec<u8>> {
    // Per the FFI method's docs, "32768 bytes is always enough".
    let mut dict = Vec::with_capacity(32_768);
    let mut dict_length = 0;
    // SAFETY: When `deflateGetDictionary` returns `Z_OK`, it holds that:
    // 1. `dict_length` elements were initialized.
    // 2. `dict_length` <= the capacity (32_768)
    // which makes `set_len` safe to call.
    unsafe {
        // Make the FFI call...
        let r = deflateGetDictionary(self.strm, dict.as_mut_ptr(), &mut dict_length);
        if r == Z_OK {
            // ...and update the length to what was initialized.
            dict.set_len(dict_length);
            Some(dict)
        } else {
            None
        }
    }
}

While the following example is sound, there is a memory leak since the inner vectors were not freed prior to the set_len call:

use rune::alloc::try_vec;

let mut vec = try_vec![try_vec![1, 0, 0],
                       try_vec![0, 1, 0],
                       try_vec![0, 0, 1]];
// SAFETY:
// 1. `old_len..0` is empty so no elements need to be initialized.
// 2. `0 <= capacity` always holds whatever `capacity` is.
unsafe {
    vec.set_len(0);
}

Normally, here, one would use clear instead to correctly drop the contents and thus not leak memory.

pub fn swap_remove(&mut self, index: usize) -> T

Removes an element from the vector and returns it.

The removed element is replaced by the last element of the vector.

This does not preserve ordering, but is O(1). If you need to preserve the element order, use remove instead.

Panics

Panics if index is out of bounds.

Examples

use rune::alloc::try_vec;

let mut v = try_vec!["foo", "bar", "baz", "qux"];

assert_eq!(v.swap_remove(1), "bar");
assert_eq!(v, ["foo", "qux", "baz"]);

assert_eq!(v.swap_remove(0), "foo");
assert_eq!(v, ["baz", "qux"]);

pub fn try_insert(&mut self, index: usize, element: T) -> Result<(), Error>

Inserts an element at position index within the vector, shifting all elements after it to the right.

Panics

Panics if index > len.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 3];
vec.try_insert(1, 4)?;
assert_eq!(vec, [1, 4, 2, 3]);
vec.try_insert(4, 5)?;
assert_eq!(vec, [1, 4, 2, 3, 5]);

pub fn remove(&mut self, index: usize) -> T

Removes and returns the element at position index within the vector, shifting all elements after it to the left.

Note: Because this shifts over the remaining elements, it has a worst-case performance of O(n). If you don’t need the order of elements to be preserved, use swap_remove instead. If you’d like to remove elements from the beginning of the Vec, consider using VecDeque::pop_front instead.

Panics

Panics if index is out of bounds.

Examples

use rune::alloc::try_vec;

let mut v = try_vec![1, 2, 3];
assert_eq!(v.remove(1), 2);
assert_eq!(v, [1, 3]);

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all elements e for which f(&e) returns false. This method operates in place, visiting each element exactly once in the original order, and preserves the order of the retained elements.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 3, 4];
vec.retain(|&x| x % 2 == 0);
assert_eq!(vec, [2, 4]);

Because the elements are visited exactly once in the original order, external state may be used to decide which elements to keep.

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 3, 4, 5];
let keep = [false, true, true, false, true];
let mut iter = keep.iter();
vec.retain(|_| *iter.next().unwrap());
assert_eq!(vec, [2, 3, 5]);

pub fn retain_mut<F>(&mut self, f: F)
where F: FnMut(&mut T) -> bool,

Retains only the elements specified by the predicate, passing a mutable reference to it.

In other words, remove all elements e such that f(&mut e) returns false. This method operates in place, visiting each element exactly once in the original order, and preserves the order of the retained elements.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 3, 4];
vec.retain_mut(|x| if *x <= 3 {
    *x += 1;
    true
} else {
    false
});
assert_eq!(vec, [2, 3, 4]);

pub fn dedup_by_key<F, K>(&mut self, key: F)
where F: FnMut(&mut T) -> K, K: PartialEq,

Removes all but the first of consecutive elements in the vector that resolve to the same key.

If the vector is sorted, this removes all duplicates.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![10, 20, 21, 30, 20];
vec.dedup_by_key(|i| *i / 10);
assert_eq!(vec, [10, 20, 30, 20]);

pub fn dedup_by<F>(&mut self, same_bucket: F)
where F: FnMut(&mut T, &mut T) -> bool,

Removes all but the first of consecutive elements in the vector satisfying a given equality relation.

The same_bucket function is passed references to two elements from the vector and must determine if the elements compare equal. The elements are passed in opposite order from their order in the slice, so if same_bucket(a, b) returns true, a is removed.

If the vector is sorted, this removes all duplicates.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec!["foo", "bar", "Bar", "baz", "bar"];
vec.dedup_by(|a, b| a.eq_ignore_ascii_case(b));
assert_eq!(vec, ["foo", "bar", "baz", "bar"]);

pub fn try_push(&mut self, value: T) -> Result<(), Error>

Appends an element to the back of a collection.

Panics

Panics if the new capacity exceeds isize::MAX bytes.

Examples

use rune::alloc::Vec;
use rune::alloc::alloc::Global;

let mut vec: Vec<i32> = Vec::try_with_capacity_in(2, Global)?;
vec.try_push(1)?;
vec.try_push(2)?;
vec.try_push(3)?;
assert_eq!(vec, [1, 2, 3]);

pub fn push_within_capacity(&mut self, value: T) -> Result<(), T>

Appends an element if there is sufficient spare capacity, otherwise an error is returned with the element.

Unlike try_push this method will not reallocate when there’s insufficient capacity. The caller should use try_reserve to ensure that there is enough capacity.

Examples

A manual, alternative to TryFromIteratorIn:

use rune::alloc::{Vec, Error};
use rune::alloc::prelude::*;

fn from_iter_fallible<T>(iter: impl Iterator<Item=T>) -> Result<Vec<T>, Error> {
    let mut vec = Vec::new();

    for value in iter {
        if let Err(value) = vec.push_within_capacity(value) {
            vec.try_reserve(1)?;
            // this cannot fail, the previous line either returned or added at least 1 free slot
            let _ = vec.push_within_capacity(value);
        }
    }

    Ok(vec)
}

assert_eq!(from_iter_fallible(0..100), Ok(Vec::try_from_iter(0..100)?));

pub fn pop(&mut self) -> Option<T>

Removes the last element from a vector and returns it, or None if it is empty.

Examples

use rune::alloc::Vec;
use rune::alloc::prelude::*;

let mut vec = Vec::try_from_iter([1, 2, 3])?;
assert_eq!(vec.pop(), Some(3));
assert_eq!(vec, [1, 2]);

pub fn try_append(&mut self, other: &mut Vec<T, A>) -> Result<(), Error>

Moves all the elements of other into self, leaving other empty.

Panics

Panics if the new capacity exceeds isize::MAX bytes.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 3];
let mut vec2 = try_vec![4, 5, 6];
vec.try_append(&mut vec2)?;
assert_eq!(vec, [1, 2, 3, 4, 5, 6]);
assert_eq!(vec2, []);

pub unsafe fn raw_iter(&self) -> RawIter<T>

Construct a raw iterator over the current vector

Safety

The caller must ensure that any pointers returned by the iterator are not dereferenced unless the object they were constructed from is still alive.

pub unsafe fn raw_iter_mut(&mut self) -> RawIterMut<T>

Construct a raw mutable iterator over the current vector

Safety

The caller must ensure that any pointers returned by the iterator are not dereferenced unless the object they were constructed from is still alive.

As a mutable iterator, this also implies that no other mutable accesses are performed over the collection this was constructed from until the returned iterator has been dropped.

pub fn drain<R>(&mut self, range: R) -> Drain<'_, T, A>
where R: RangeBounds<usize>,

Removes the specified range from the vector in bulk, returning all removed elements as an iterator. If the iterator is dropped before being fully consumed, it drops the remaining removed elements.

The returned iterator keeps a mutable borrow on the vector to optimize its implementation.

Panics

Panics if the starting point is greater than the end point or if the end point is greater than the length of the vector.

Leaking

If the returned iterator goes out of scope without being dropped (due to mem::forget, for example), the vector may have lost and leaked elements arbitrarily, including elements outside the range.

Examples

use rune::alloc::{try_vec, Vec};
use rune::alloc::prelude::*;

let mut v = try_vec![1, 2, 3];
let u: Vec<_> = v.drain(1..).try_collect()?;
assert_eq!(v, &[1]);
assert_eq!(u, &[2, 3]);

// A full range clears the vector, like `clear()` does
v.drain(..);
assert_eq!(v, &[]);

pub fn clear(&mut self)

Clears the vector, removing all values.

Note that this method has no effect on the allocated capacity of the vector.

Examples

use rune::alloc::try_vec;

let mut v = try_vec![1, 2, 3];
v.clear();
assert!(v.is_empty());

pub const fn len(&self) -> usize

Returns the number of elements in the vector, also referred to as its ‘length’.

Examples

use rune::alloc::Vec;
use rune::alloc::alloc::Global;

let mut a = Vec::new_in(Global);

for value in 0..3 {
    a.try_push(value)?;
}

assert_eq!(a.len(), 3);

pub fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

Examples

use rune::alloc::Vec;

let mut v = Vec::new();
assert!(v.is_empty());

v.try_push(1)?;
assert!(!v.is_empty());

pub fn try_split_off(&mut self, at: usize) -> Result<Vec<T, A>, Error>
where A: Clone,

Splits the collection into two at the given index.

Returns a newly allocated vector containing the elements in the range [at, len). After the call, the original vector will be left containing the elements [0, at) with its previous capacity unchanged.

Panics

Panics if at > len.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 3];
let vec2 = vec.try_split_off(1)?;
assert_eq!(vec, [1]);
assert_eq!(vec2, [2, 3]);

pub fn try_resize_with<F>(&mut self, new_len: usize, f: F) -> Result<(), Error>
where F: FnMut() -> T,

Resizes the Vec in-place so that len is equal to new_len.

If new_len is greater than len, the Vec is extended by the difference, with each additional slot filled with the result of calling the closure f. The return values from f will end up in the Vec in the order they have been generated.

If new_len is less than len, the Vec is simply truncated.

This method uses a closure to create new values on every push. If you’d rather Clone a given value, use Vec::try_resize. If you want to use the Default trait to generate values, you can pass Default::default as the second argument.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 3];
vec.try_resize_with(5, Default::default)?;
assert_eq!(vec, [1, 2, 3, 0, 0]);

let mut vec = try_vec![];
let mut p = 1;
vec.try_resize_with(4, || { p *= 2; p })?;
assert_eq!(vec, [2, 4, 8, 16]);

pub fn leak<'a>(self) -> &'a mut [T]
where A: 'a,

Consumes and leaks the Vec, returning a mutable reference to the contents, &'a mut [T]. Note that the type T must outlive the chosen lifetime 'a. If the type has only static references, or none at all, then this may be chosen to be 'static.

As of Rust 1.57, this method does not reallocate or shrink the Vec, so the leaked allocation may include unused capacity that is not part of the returned slice.

This function is mainly useful for data that lives for the remainder of the program’s life. Dropping the returned reference will cause a memory leak.

Examples

Simple usage:

use rune::alloc::try_vec;

let x = try_vec![1, 2, 3];
let static_ref: &'static mut [usize] = x.leak();
static_ref[0] += 1;
assert_eq!(static_ref, &[2, 2, 3]);

pub fn spare_capacity_mut(&mut self) -> &mut [MaybeUninit<T>]

Returns the remaining spare capacity of the vector as a slice of MaybeUninit<T>.

The returned slice can be used to fill the vector with data (e.g. by reading from a file) before marking the data as initialized using the set_len method.

Examples

use rune::alloc::Vec;

// Allocate vector big enough for 10 elements.
let mut v = Vec::try_with_capacity(10)?;

// Fill in the first 3 elements.
let uninit = v.spare_capacity_mut();
uninit[0].write(0);
uninit[1].write(1);
uninit[2].write(2);

// Mark the first 3 elements of the vector as being initialized.
unsafe {
    v.set_len(3);
}

assert_eq!(&v, &[0, 1, 2]);

pub fn split_at_spare_mut(&mut self) -> (&mut [T], &mut [MaybeUninit<T>])

Returns vector content as a slice of T, along with the remaining spare capacity of the vector as a slice of MaybeUninit<T>.

The returned spare capacity slice can be used to fill the vector with data (e.g. by reading from a file) before marking the data as initialized using the set_len method.

Note that this is a low-level API, which should be used with care for optimization purposes. If you need to append data to a Vec you can use try_push, try_extend, try_extend_from_slice, try_extend_from_within, try_insert, try_append, try_resize or try_resize_with, depending on your exact needs.

Examples

use rune::alloc::try_vec;

let mut v = try_vec![1, 1, 2];

// Reserve additional space big enough for 10 elements.
v.try_reserve(10)?;

let (init, uninit) = v.split_at_spare_mut();
let sum = init.iter().copied().sum::<u32>();

// Fill in the next 4 elements.
uninit[0].write(sum);
uninit[1].write(sum * 2);
uninit[2].write(sum * 3);
uninit[3].write(sum * 4);

// Mark the 4 elements of the vector as being initialized.
unsafe {
    let len = v.len();
    v.set_len(len + 4);
}

assert_eq!(&v, &[1, 1, 2, 4, 8, 12, 16]);

impl<T, A> Vec<T, A>
where A: Allocator, T: TryClone,

pub fn try_resize(&mut self, new_len: usize, value: T) -> Result<(), Error>

Resizes the Vec in-place so that len is equal to new_len.

If new_len is greater than len, the Vec is extended by the difference, with each additional slot filled with value. If new_len is less than len, the Vec is simply truncated.

This method requires T to implement Clone, in order to be able to clone the passed value. If you need more flexibility (or want to rely on Default instead of Clone), use Vec::try_resize_with. If you only need to resize to a smaller size, use Vec::truncate.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec!["hello"];
vec.try_resize(3, "world")?;
assert_eq!(vec, ["hello", "world", "world"]);

let mut vec = try_vec![1, 2, 3, 4];
vec.try_resize(2, 0)?;
assert_eq!(vec, [1, 2]);

pub fn try_extend_from_slice(&mut self, other: &[T]) -> Result<(), Error>

Clones and appends all elements in a slice to the Vec.

Iterates over the slice other, clones each element, and then appends it to this Vec. The other slice is traversed in-order.

Note that this function is same as try_extend except that it is specialized to work with slices instead. If and when Rust gets specialization this function will likely be deprecated (but still available).

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![1];
vec.try_extend_from_slice(&[2, 3, 4]);
assert_eq!(vec, [1, 2, 3, 4]);

pub fn try_extend_from_within<R>(&mut self, src: R) -> Result<(), Error>
where R: RangeBounds<usize>,

Copies elements from src range to the end of the vector.

Panics

Panics if the starting point is greater than the end point or if the end point is greater than the length of the vector.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![0, 1, 2, 3, 4];

vec.try_extend_from_within(2..);
assert_eq!(vec, [0, 1, 2, 3, 4, 2, 3, 4]);

vec.try_extend_from_within(..2);
assert_eq!(vec, [0, 1, 2, 3, 4, 2, 3, 4, 0, 1]);

vec.try_extend_from_within(4..8);
assert_eq!(vec, [0, 1, 2, 3, 4, 2, 3, 4, 0, 1, 4, 2, 3, 4]);

impl<T, A, const N: usize> Vec<[T; N], A>
where A: Allocator,

pub fn into_flattened(self) -> Vec<T, A>

Takes a Vec<[T; N]> and flattens it into a Vec<T>.

Panics

Panics if the length of the resulting vector would overflow a usize.

This is only possible when flattening a vector of arrays of zero-sized types, and thus tends to be irrelevant in practice. If size_of::<T>() > 0, this will never panic.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![[1, 2, 3], [4, 5, 6], [7, 8, 9]];
assert_eq!(vec.pop(), Some([7, 8, 9]));

let mut flattened = vec.into_flattened();
assert_eq!(flattened.pop(), Some(6));

impl<T, A> Vec<T, A>
where A: Allocator, T: PartialEq,

pub fn dedup(&mut self)

Removes consecutive repeated elements in the vector according to the PartialEq trait implementation.

If the vector is sorted, this removes all duplicates.

Examples

use rune::alloc::try_vec;

let mut vec = try_vec![1, 2, 2, 3, 2];
vec.dedup();
assert_eq!(vec, [1, 2, 3, 2]);