rune

Module compile

The Rune compiler.

The main entry to compiling rune source is prepare which uses this compiler. In here you’ll just find compiler-specific types.

Re-exports

pub use self::meta::MetaRef;

pub use self::meta::SourceMeta;

Modules

meta

Compiler metadata for Rune.

Structs

Context

Context used for the Rune language.

Error

An error raised by the compiler.

FileSourceLoader

A filesystem-based source loader.

ImportStep

A single step in an import.

ItemId

The identifier of an item.

Location

A fully descriptive location which is a combination of a SourceId and a Span.

MetaError

A meta error.

MetaInfo

Provides an owned human-readable description of a meta item.

NoopSourceLoader

A source loader which does not support loading anything and will error.

Options

Options that can be provided to the compiler.

ParseOptionError

Error raised when trying to parse an invalid option.

Enums

ContextError

An error raised when building the context.

LinkerError

Errors that can be raised when linking units.

Traits

CompileVisitor

A visitor that will be called for every language item compiled.

Located

Trait for things that have a Location.

Named

The trait used for something that can be statically named.

SourceLoader

A source loader.

Type Aliases

Result

Helper alias for compile results.