Enum rune::compile::meta::Kind

#[non_exhaustive]
pub enum Kind {

    Type {
        parameters: Hash,
    },
    Struct {
        fields: Fields,
        constructor: Option<Signature>,
        parameters: Hash,
    },
    Variant {
        enum_hash: Hash,
        index: usize,
        fields: Fields,
        constructor: Option<Signature>,
    },
    Enum {
        parameters: Hash,
    },
    Macro,
    AttributeMacro,
    Function {
        associated: Option<AssociatedKind>,
        signature: Signature,
        is_test: bool,
        is_bench: bool,
        parameters: Hash,
        container: Option<Hash>,
        parameter_types: Vec<Hash>,
    },
    Closure {
        call: Call,
        do_move: bool,
    },
    AsyncBlock {
        call: Call,
        do_move: bool,
    },
    Const,
    ConstFn {
        id: NonZeroId,
    },
    Import(Import),
    Module,

}

Compile-time metadata kind about a unit.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Type

Fields

parameters: Hash

Hash of generic parameters.

The type is completely opaque. We have no idea about what it is with the exception of it having a type hash.

Struct

Fields

fields: Fields

Fields information.

constructor: Option<Signature>

Native constructor for this struct.

parameters: Hash

Hash of generic parameters.

Metadata about a struct.

Variant

Fields

enum_hash: Hash

Type hash of the enum this unit variant belongs to.

index: usize

The index of the variant.

fields: Fields

Fields information.

constructor: Option<Signature>

Native constructor for this variant.

Metadata about an empty variant.

Enum

Fields

parameters: Hash

Hash of generic parameters.

An enum item.

Macro

A macro item.

AttributeMacro

An attribute macro item.

Function

Fields

associated: Option<AssociatedKind>

The associated kind of the function, if it is an associated function.

signature: Signature

Native signature for this function.

is_test: bool

Whether this function has a #[test] annotation

is_bench: bool

Whether this function has a #[bench] annotation.

parameters: Hash

Hash of generic parameters.

container: Option<Hash>

The container of the associated function.

parameter_types: Vec<Hash>

Parameter types.

A function declaration.

Closure

Fields

call: Call

Runtime calling convention.

do_move: bool

If the closure moves its environment.

A closure.

AsyncBlock

Fields

call: Call

Runtime calling convention.

do_move: bool

If the async block moves its environment.

An async block.

Const

The constant expression.

ConstFn

Fields

id: NonZeroId

Opaque identifier for the constant function.

A constant function.

Import(Import)

Purely an import.

Module

A module.