Struct rune::compile::ItemBuf

pub struct ItemBuf<A = Global>
where
    A: Allocator,
{ /* private fields */ }

The name of an item in the Rune Language.

This is made up of a collection of strings, like ["foo", "bar"]. This is indicated in rune as foo::bar.

An item can also belongs to a crate, which in rune could be indicated as ::crate::foo::bar. These items must be constructed using ItemBuf::with_crate.

Items are inlined if they are smaller than 32 bytes.

Panics

The max length of a string component is is 2**14 = 16384. Attempting to add a string larger than that will panic. This also constitutes the maximum number of nested sibling components that can exist in a single source file since they all use anonymous identifiers.

Component encoding

The following details internal implementation details of an Item, and is not exposed through its API. It is provided here in case you need to work with the internal of an item.

A single component is encoded as:

A two byte tag as a u16 in native endianess, indicating its type (least significant 2 bits) and data (most significant 14 bits).

If the type is a STRING, the data is treated as the length of the string. Any other type this the data is treated as the numeric id of the component.

If the type is a STRING, the tag is repeated at the end of it to allow for seeking backwards. This is not the case for other types. Since they are fixed size its not necessary.

So all in all, a string is encoded as this where the d part indicates the length of the string:

dddddddd ddddddtt *string content* dddddddd ddddddtt

And any other component is just the two bytes where the d part makes up a numerical component:

dddddddd ddddddtt

Implementations

impl<A> ItemBuf<A>
where A: Allocator,

pub fn push<C>(&mut self, c: C) -> Result<(), Error>
where C: IntoComponent,

Push the given component to the current item.

pub fn pop(&mut self) -> Result<Option<Component>, Error>

Push the given component to the current item.

pub fn extend<I>(&mut self, i: I) -> Result<(), Error>
where I: IntoIterator, <I as IntoIterator>::Item: IntoComponent,

Extend the current item with an iterator.

pub fn clear(&mut self)

Clear the current item.

impl ItemBuf

pub const fn new() -> ItemBuf

Construct a new empty item.

Examples

use rune::compile::ItemBuf;

let item = ItemBuf::new();
let mut it = item.iter();

assert_eq!(it.next(), None);

pub fn with_item<I>(iter: I) -> Result<ItemBuf, Error>
where I: IntoIterator, <I as IntoIterator>::Item: IntoComponent,

Construct a new item with the given path.

Examples

use rune::compile::{ComponentRef, ItemBuf};

let item = ItemBuf::with_item(["foo", "bar"])?;
let mut it = item.iter();

assert_eq!(it.next(), Some(ComponentRef::Str("foo")));
assert_eq!(it.next(), Some(ComponentRef::Str("bar")));
assert_eq!(it.next(), None);

pub fn with_crate(name: &str) -> Result<ItemBuf, Error>

Construct item for a crate.

Examples

use rune::compile::{ComponentRef, ItemBuf};

let mut item = ItemBuf::with_crate("std")?;
item.push("foo");
assert_eq!(item.as_crate(), Some("std"));

let mut it = item.iter();
assert_eq!(it.next(), Some(ComponentRef::Crate("std")));
assert_eq!(it.next(), Some(ComponentRef::Str("foo")));
assert_eq!(it.next(), None);

pub fn with_crate_item<I>(name: &str, iter: I) -> Result<ItemBuf, Error>
where I: IntoIterator, <I as IntoIterator>::Item: IntoComponent,

Create a crated item with the given name.

Examples

use rune::compile::{ComponentRef, ItemBuf};

let item = ItemBuf::with_crate_item("std", ["option"])?;
assert_eq!(item.as_crate(), Some("std"));

let mut it = item.iter();
assert_eq!(it.next(), Some(ComponentRef::Crate("std")));
assert_eq!(it.next(), Some(ComponentRef::Str("option")));
assert_eq!(it.next(), None);