rune::compile

Trait Named

pub trait Named {
    const ITEM: &'static Item;

    // Provided methods
    fn full_name(f: &mut Formatter<'_>) -> Result { ... }
    fn display() -> impl Display { ... }
}

The trait used for something that can be statically named.

Required Associated Constants

const ITEM: &'static Item

The name item.

Provided Methods

fn full_name(f: &mut Formatter<'_>) -> Result

The full name of the type.

fn display() -> impl Display

Return a display wrapper for the full name of the type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.