rune

Module hash

Utilities for working with hashes.

Structs

Hash

The primitive hash that among other things is used to reference items, types, and native functions.

ParametersBuilder

Helper to build a parameters hash.

TooManyParameters

Traits

IntoHash

Trait for types which can be converted into a Hash.

ToTypeHash

Helper trait used to convert a type into a type hash.