rune

Module languageserver

Available on crate feature languageserver only.

Utility for building a language server.

Modules

envelope

Types to deserialize.

Structs

Builder

A builder for a language server.

LanguageServer

The instance of a language server, as constructed through builder().

Unset

Unset placeholder I/O types for language server.

Functions

builder

Construct a new empty builder without any configured I/O.