Macro rune::static_env

macro_rules! static_env {
    () => { ... };
}

Defines a static budget and environment implementation suitable for singlethreaded no-std environments. This can be used in #[no_std] environments to implement the necessary hooks for Rune to work.

The alternative is to implement these manually.

If the std feature is enabled, thread-local budgeting will be used and calling this will do nothing.

Examples

rune::no_std::static_env!();