Struct rune::module::TypeMut

pub struct TypeMut<'a, T>
where
    T: ?Sized + TypeOf,
{ /* private fields */ }

Handle to a a type inserted into a module which allows for mutation of its metadata.

This is returned by the following methods:

Module::ty - after a type has been inserted.

Module::type_meta - to modify type metadata for an already inserted type.

Implementations

impl<'a, T> TypeMut<'a, T>
where T: ?Sized + TypeOf,

pub fn docs<I>(self, docs: I) -> Result<Self, ContextError>
where I: IntoIterator, I::Item: AsRef<str>,

Set documentation for an inserted type.

This completely replaces any existing documentation.

pub fn static_docs( self, docs: &'static [&'static str] ) -> Result<Self, ContextError>

Set static documentation.

This completely replaces any existing documentation.

pub fn deprecated<S>(self, deprecated: S) -> Result<Self, ContextError>
where S: AsRef<str>,

Mark the given type as deprecated.

pub fn make_named_struct( self, fields: &'static [&'static str] ) -> Result<Self, ContextError>

Mark the current type as a struct with named fields.

pub fn make_unnamed_struct(self, fields: usize) -> Result<Self, ContextError>

Mark the current type as a struct with unnamed fields.

pub fn make_empty_struct(self) -> Result<Self, ContextError>

Mark the current type as an empty struct.

pub fn make_enum( self, variants: &'static [&'static str] ) -> Result<EnumMut<'a, T>, ContextError>

Mark the current type as an enum.

pub fn constructor<F, A>(self, constructor: F) -> Result<Self, ContextError>
where F: Function<A, Plain, Return = T>,

Register a constructor method for the current type.