rune::module
Type Alias ModuleMeta

pub type ModuleMeta = fn() -> Result<ModuleMetaData>;

Type used to collect and store module metadata through the #[rune::module] macro.

This is the argument type for Module::from_meta, and is from a public API perspective completely opaque and might change for any release.

Calling and making use of ModuleMeta manually despite this warning might lead to future breakage.