Module rune::modules

Public packages that can be used to provide extract functionality to virtual machines.

These are usually included through Context::with_default_modules.

Modules

any

Dynamic typing and type reflection.

bytes

The bytes module.

capture_io

I/O module capable of capturing what’s been written to a buffer.

char

The character module for Rune.

clone

The cloning trait for Rune.

cmp

Comparison and ordering.

collections

Dynamic collections.

core

Core types and methods in Rune.

disable_io

I/O methods which will cause any output to be ignored.

f64

Floating point numbers.

fmt

Formatting text.

future

Asynchronous computations.

generator

Generators.

hash

Hashing types.

i64

Integers.

io

I/O functions.

iter

Iterators.

macros

Macro support.

mem

Working with memory.

num

Working with numbers.

object

The dynamic Object container.

ops

Overloadable operators and associated types.

option

The Option type.

result

The Result type.

stream

Asynchronous streams.

string

Strings.

test

Testing and benchmarking.

tuple

The Tuple fixed collection.

vec

The Vec dynamic vector.