Module rune::runtime

Runtime module for Rune.

Re-exports

pub use self::debug::DebugInfo;

pub use self::debug::DebugInst;

pub use self::format::Format;

pub use self::format::FormatSpec;

pub use self::unit::Unit;

pub use self::unit::UnitStorage;

pub use self::value::TypeValue;

pub use self::vm_error::try_result;

pub use self::vm_error::TryFromResult;

Modules

budget

Budgeting module for Runestick.

debug

Debug information for units.

format

Types for dealing with formatting specifications.

unit

A single execution unit in the rune virtual machine.

Structs

AccessError

An error raised while downcasting.

AnyObj

Our own private dynamic Any implementation.

AnyObjVtable

The vtable for any type stored in the virtual machine.

AnyTypeInfo

Type information for the Any type.

BorrowMut

Guard for data exclusively borrowed from a slot in the virtual machine.

BorrowRef

Guard for a data borrowed from a slot in the virtual machine.

Bytes

A vector of bytes.

CallFrame

A call frame.

DebugLabel

A label that can be jumped to.

EmptyStruct

A empty with a well-defined type.

Formatter

A formatter for the rune virtual machine.

FullTypeOf

Full type information.

Function

The type of a function in Rune.

Future

A type-erased future that can only be unsafely polled in combination with the virtual machine that created it.

Generator

The return value of a function producing a generator.

Hasher

The default hasher used in Rune.

Iterator

An owning iterator.

Mut

A strong mutable reference to the given type.

Object

Struct representing a dynamic anonymous object.

OwnedTuple

Struct representing a dynamic anonymous object.

Protocol

A built in instance function.

Range

Type for a range expression start..end.

RangeFrom

Type for a from range expression start...

RangeFull

Type for a full range expression ...

RangeInclusive

Type for an inclusive range expression start..=end.

RangeTo

Type for an inclusive range expression ..end.

RangeToInclusive

Type for an inclusive range expression ..=end.

RawMut

A raw guard to a Ref.

RawRef

A raw guard to a Ref.

RawStr

A raw static string.

Ref

A strong reference to the given type.

Rtti

Runtime information on variant.

RuntimeContext

Static run context visible to the virtual machine.

RuntimeError

An opaque simple runtime error.

SharedPointerGuard

A drop guard for a shared value.

Snapshot

Snapshot that can be used to indicate how the value was being accessed at the time of an error.

Stack

The stack of the virtual machine, where all values are stored.

StackError

An error raised when interacting with the stack.

StaticString

Struct representing a static string.

StaticType

Static type information.

Stream

A stream with a stored virtual machine.

Struct

An object with a well-defined type.

SyncFunction

A callable sync function. This currently only supports a subset of values that are supported by the Vm.

Tuple

The type of a tuple slice.

TupleStruct

A tuple with a well-defined type.

Type

A value representing a type in the virtual machine.

Value

An entry on the stack.

Variant

The variant of a type.

VariantRtti

Runtime information on variant.

Vec

Struct representing a dynamic vector.

VecTuple

A helper type to deserialize arrays with different interior types.

Vm

A stack which references variables indirectly from a slab.

VmError

A virtual machine error which includes tracing information.

VmExecution

The execution environment for a virtual machine.

VmIntegerRepr

A type-erased rust number.

VmSendExecution

A wrapper that makes VmExecution Send.

Enums

AnyObjError

Errors raised during casting operations.

Call

The calling convention of a function.

ConstValue

A constant value.

ControlFlow

Used to tell an operation whether it should exit early or go on as usual.

GeneratorState

The state of a generator.

Inst

An operation in the stack-based virtual machine.

InstAddress

How an instruction addresses a value.

InstAssignOp

An operation between two values on the machine.

InstOp

An operation between two values on the machine.

InstRange

Range limits of a range expression.

InstTarget

The target of an operation.

InstValue

A literal value that can be pushed.

InstVariant

A variant that can be constructed.

PanicReason

Pre-canned panic reasons.

TypeCheck

Type checks for built-in types.

TypeInfo

Type information about a value, that can be printed for human consumption through its Display implementation.

VariantData

The data of the variant.

VmResult

A result produced by the virtual machine.

Traits

Args

Trait for converting arguments onto the stack.

FromValue

Trait for converting types from the dynamic Value container.

GuardedArgs

Trait for converting arguments onto the stack.

IteratorTrait

The trait for interacting with an iterator.

MaybeTypeOf

A type that might or might not have a concrete type.

ToValue

Trait for converting types into the dynamic Value container.

TypeOf

Trait used for Rust types for which we can determine the runtime type of.

UnsafeFromValueDeprecated

A potentially unsafe conversion for value conversion.

UnsafeToMut

Unsafe to mut coercion.

UnsafeToRef

Unsafe to ref coercion.

UnsafeToValue

Trait for converting types into values.

Functions

from_value

Convert something into the dynamic Value.

to_value

Convert something into the dynamic Value.

Derive Macros

FromValue

Derive macro for the FromValue trait for converting types from the dynamic Value container.

ToValue

Derive macro for the ToValue trait for converting types into the dynamic Value container.