Enum rune::runtime::ConstValue

pub enum ConstValue {
    EmptyTuple,
    Byte(u8),
    Char(char),
    Bool(bool),
    Integer(i64),
    Float(f64),
    String(String),
    Bytes(Bytes),
    Vec(Vec<ConstValue>),
    Tuple(Box<[ConstValue]>),
    Object(HashMap<String, ConstValue>),
    Option(Option<Box<ConstValue>>),
}

A constant value.

Variants

EmptyTuple

A constant unit.

Byte(u8)

A byte.

Char(char)

A character.

Bool(bool)

A boolean constant value.

Integer(i64)

An integer constant.

Float(f64)

An float constant.

String(String)

A string constant designated by its slot.

Bytes(Bytes)

A byte string.

Vec(Vec<ConstValue>)

A vector of values.

Tuple(Box<[ConstValue]>)

An anonymous tuple.

Object(HashMap<String, ConstValue>)

An anonymous object.

Option(Option<Box<ConstValue>>)

An option.

Implementations

impl ConstValue

pub fn as_value(&self) -> Result<Value>

Convert into virtual machine value.

We provide this associated method since a constant value can be converted into a value infallibly, which is not captured by the trait otherwise.

pub fn into_bool(self) -> Result<bool, Self>

Try to coerce into boolean.

pub fn type_info(&self) -> TypeInfo

Get the type information of the value.