Enum rune::runtime::ControlFlow

pub enum ControlFlow {
    Continue(Value),
    Break(Value),
}

Used to tell an operation whether it should exit early or go on as usual.

This acts as the basis of the TRY protocol in Rune.

Examples

use std::ops::ControlFlow;

let c = ControlFlow::Continue(42);
assert_eq!(c.0, 42);
assert_eq!(c, ControlFlow::Continue(42));

Variants

Continue(Value)

Move on to the next phase of the operation as normal.

Break(Value)

Exit the operation without running subsequent phases.