Enum rune::runtime::Value

pub enum Value {

    Bool(bool),
    Byte(u8),
    Char(char),
    Integer(i64),
    Float(f64),
    Type(Type),
    Ordering(Ordering),
    String(Shared<String>),
    Bytes(Shared<Bytes>),
    Vec(Shared<Vec>),
    EmptyTuple,
    Tuple(Shared<OwnedTuple>),
    Object(Shared<Object>),
    RangeFrom(Shared<RangeFrom>),
    RangeFull(Shared<RangeFull>),
    RangeInclusive(Shared<RangeInclusive>),
    RangeToInclusive(Shared<RangeToInclusive>),
    RangeTo(Shared<RangeTo>),
    Range(Shared<Range>),
    ControlFlow(Shared<ControlFlow>),
    Future(Shared<Future>),
    Stream(Shared<Stream<Vm>>),
    Generator(Shared<Generator<Vm>>),
    GeneratorState(Shared<GeneratorState>),
    Option(Shared<Option<Value>>),
    Result(Shared<Result<Value, Value>>),
    EmptyStruct(Shared<EmptyStruct>),
    TupleStruct(Shared<TupleStruct>),
    Struct(Shared<Struct>),
    Variant(Shared<Variant>),
    Function(Shared<Function>),
    Format(Shared<Format>),
    Iterator(Shared<Iterator>),
    Any(Shared<AnyObj>),

}

An entry on the stack.

Variants

Bool(bool)

A boolean.

Byte(u8)

A single byte.

Char(char)

A character.

Integer(i64)

A number.

Float(f64)

A float.

Type(Type)

A type hash. Describes a type in the virtual machine.

Ordering(Ordering)

Ordering.

String(Shared<String>)

A UTF-8 string.

Bytes(Shared<Bytes>)

A byte string.

Vec(Shared<Vec>)

A vector containing any values.

EmptyTuple

The unit value.

Tuple(Shared<OwnedTuple>)

A tuple.

Object(Shared<Object>)

An object.

RangeFrom(Shared<RangeFrom>)

A range start..

RangeFull(Shared<RangeFull>)

A full range ..

RangeInclusive(Shared<RangeInclusive>)

A full range start..=end

RangeToInclusive(Shared<RangeToInclusive>)

A full range ..=end

RangeTo(Shared<RangeTo>)

A full range ..end

Range(Shared<Range>)

A range start..end.

ControlFlow(Shared<ControlFlow>)

A control flow indicator.

Future(Shared<Future>)

A stored future.

Stream(Shared<Stream<Vm>>)

A Stream.

Generator(Shared<Generator<Vm>>)

A stored generator.

GeneratorState(Shared<GeneratorState>)

Generator state.

Option(Shared<Option<Value>>)

An empty value indicating nothing.

Result(Shared<Result<Value, Value>>)

A stored result in a slot.

EmptyStruct(Shared<EmptyStruct>)

An struct with a well-defined type.

TupleStruct(Shared<TupleStruct>)

A tuple with a well-defined type.

Struct(Shared<Struct>)

An struct with a well-defined type.

Variant(Shared<Variant>)

The variant of an enum.

Function(Shared<Function>)

A stored function pointer.

Format(Shared<Format>)

A value being formatted.

Iterator(Shared<Iterator>)

An iterator.

Any(Shared<AnyObj>)

An opaque value that can be downcasted.

Implementations

impl Value

pub fn string_display(&self, f: &mut Formatter) -> VmResult<()>

Format the value using the Protocol::STRING_DISPLAY protocol.

Requires a work buffer buf which will be used in case the value provided requires out-of-line formatting. This must be cleared between calls and can be re-used.

You must use Vm::with to specify which virtual machine this function is called inside.

Panics

This function will panic if called outside of a virtual machine.

pub fn string_debug(&self, f: &mut Formatter) -> VmResult<()>

Debug format the value using the STRING_DEBUG protocol.

You must use Vm::with to specify which virtual machine this function is called inside.

Panics

This function will panic if called outside of a virtual machine.

pub fn into_iter(self) -> VmResult<Iterator>

Convert value into an iterator using the Protocol::INTO_ITER protocol.

You must use Vm::with to specify which virtual machine this function is called inside.

Errors

This function will error if called outside of a virtual machine context.

pub fn into_future(self) -> VmResult<Shared<Future>>

Coerce into a future, or convert into a future using the Protocol::INTO_FUTURE protocol.

You must use Vm::with to specify which virtual machine this function is called inside.

Errors

This function errors in case the provided type cannot be converted into a future without the use of a Vm and one is not provided through the environment.

pub fn into_type_name(self) -> VmResult<String>

Retrieves a human readable type name for the current value.

You must use Vm::with to specify which virtual machine this function is called inside.

Errors

This function errors in case the provided type cannot be converted into a name without the use of a Vm and one is not provided through the environment.

pub fn vec(vec: Vec<Value>) -> VmResult<Self>

Construct a vector.

pub fn tuple(vec: Vec<Value>) -> VmResult<Self>

Construct a tuple.

pub fn empty_struct(rtti: Arc<Rtti>) -> VmResult<Self>

Construct an empty.

pub fn tuple_struct(rtti: Arc<Rtti>, vec: Vec<Value>) -> VmResult<Self>

Construct a typed tuple.

pub fn unit_variant(rtti: Arc<VariantRtti>) -> VmResult<Self>

Construct an empty variant.

pub fn tuple_variant(rtti: Arc<VariantRtti>, vec: Vec<Value>) -> VmResult<Self>

Construct a tuple variant.

pub fn take(self) -> VmResult<Self>

Take the interior value.

pub fn into_unit(self) -> VmResult<()>

Try to coerce value into a unit.

pub fn as_bool(&self) -> VmResult<bool>

Try to coerce value into a boolean.

pub fn into_bool(self) -> VmResult<bool>

Try to coerce value into a boolean.

pub fn as_byte(&self) -> VmResult<u8>

Try to coerce value into a byte.

pub fn into_byte(self) -> VmResult<u8>

Try to coerce value into a byte.

pub fn as_char(&self) -> VmResult<char>

Try to coerce value into a character.

pub fn into_char(self) -> VmResult<char>

Try to coerce value into a character.

pub fn as_integer(&self) -> VmResult<i64>

Try to coerce value into an integer.

pub fn into_integer(self) -> VmResult<i64>

Try to coerce value into an integer.

pub fn as_float(&self) -> VmResult<f64>

Try to coerce value into a float.

pub fn into_float(self) -> VmResult<f64>

Try to coerce value into a float.

pub fn as_type(&self) -> VmResult<Type>

Try to coerce value into a type.

pub fn into_type(self) -> VmResult<Type>

Try to coerce value into a type.

pub fn as_usize(&self) -> VmResult<usize>

Try to coerce value into a usize.

pub fn into_usize(self) -> VmResult<usize>

Try to coerce value into a usize.

pub fn as_ordering(&self) -> VmResult<Ordering>

Try to coerce value into an Ordering.

pub fn into_ordering(self) -> VmResult<Ordering>

Try to coerce value into an Ordering.

pub fn into_result(self) -> VmResult<Shared<Result<Value, Value>>>

Try to coerce value into a result.

pub fn as_result(&self) -> VmResult<&Shared<Result<Value, Value>>>

Try to coerce value into a result.

pub fn into_generator(self) -> VmResult<Shared<Generator<Vm>>>

Try to coerce value into a generator.

pub fn into_stream(self) -> VmResult<Shared<Stream<Vm>>>

Try to coerce value into a stream.

pub fn into_generator_state(self) -> VmResult<Shared<GeneratorState>>

Try to coerce value into a future.

pub fn into_option(self) -> VmResult<Shared<Option<Value>>>

Try to coerce value into an option.

pub fn into_string(self) -> VmResult<Shared<String>>

Try to coerce value into a string.

pub fn into_bytes(self) -> VmResult<Shared<Bytes>>

Try to coerce value into bytes.

pub fn into_vec(self) -> VmResult<Shared<Vec>>

Try to coerce value into a vector.

pub fn into_tuple(self) -> VmResult<Shared<OwnedTuple>>

Try to coerce value into a tuple.

pub fn into_object(self) -> VmResult<Shared<Object>>

Try to coerce value into an object.

pub fn into_range_from(self) -> VmResult<Shared<RangeFrom>>

Try to coerce value into a RangeFrom.

pub fn into_range_full(self) -> VmResult<Shared<RangeFull>>

Try to coerce value into a RangeFull.

pub fn into_range_to_inclusive(self) -> VmResult<Shared<RangeToInclusive>>

Try to coerce value into a RangeToInclusive.

pub fn into_range_inclusive(self) -> VmResult<Shared<RangeInclusive>>

Try to coerce value into a RangeInclusive.

pub fn into_range_to(self) -> VmResult<Shared<RangeTo>>

Try to coerce value into a RangeTo.

pub fn into_range(self) -> VmResult<Shared<Range>>

Try to coerce value into a Range.

pub fn into_control_flow(self) -> VmResult<Shared<ControlFlow>>

Try to coerce value into a ControlFlow.

pub fn into_function(self) -> VmResult<Shared<Function>>

Try to coerce value into a function pointer.

pub fn into_format(self) -> VmResult<Shared<Format>>

Try to coerce value into a format spec.

pub fn into_iterator(self) -> VmResult<Shared<Iterator>>

Try to coerce value into an iterator.

pub fn into_any(self) -> VmResult<Shared<AnyObj>>

Try to coerce value into an opaque value.

pub fn into_any_ptr<T>(self) -> VmResult<(NonNull<T>, RawRef)>where T: Any,

Try to coerce value into a ref and an associated guard.

Safety

This coerces a strong guard to the value into its raw components.

It is up to the caller to ensure that the returned pointer does not outlive the returned guard, not the virtual machine the value belongs to.

pub fn into_any_mut<T>(self) -> VmResult<(NonNull<T>, RawMut)>where T: Any,

Try to coerce value into a ref and an associated guard.

Safety

This coerces a strong guard to the value into its raw components.

It is up to the caller to ensure that the returned pointer does not outlive the returned guard, not the virtual machine the value belongs to.

pub fn type_hash(&self) -> Result<Hash, VmError>

Get the type hash for the current value.

One notable feature is that the type of a variant is its container enum, and not the type hash of the variant itself.

pub fn type_info(&self) -> VmResult<TypeInfo>

Get the type information for the current value.

pub fn partial_eq(a: &Value, b: &Value) -> VmResult<bool>

Perform a partial equality test between two values.

This is the basis for the eq operation (partial_eq / ‘==’).

External types will use the Protocol::PARTIAL_EQ protocol when invoked through this function.

Errors

This function will error if called outside of a virtual machine context.

pub fn hash(&self, hasher: &mut Hasher) -> VmResult<()>

Hash the current value.

pub fn eq(&self, b: &Value) -> VmResult<bool>

Perform a total equality test between two values.

This is the basis for the eq operation (==).

External types will use the Protocol::EQ protocol when invoked through this function.

Errors

This function will error if called outside of a virtual machine context.

pub fn partial_cmp(a: &Value, b: &Value) -> VmResult<Option<Ordering>>

Perform a partial ordering comparison between two values.

This is the basis for the comparison operation.

External types will use the Protocol::PARTIAL_CMP protocol when invoked through this function.

Errors

This function will error if called outside of a virtual machine context.

pub fn cmp(a: &Value, b: &Value) -> VmResult<Ordering>

Perform a total ordering comparison between two values.

This is the basis for the comparison operation (cmp).

External types will use the Protocol::CMP protocol when invoked through this function.

Errors

This function will error if called outside of a virtual machine context.