rune::runtime

Enum VmResult

pub enum VmResult<T> {
    Ok(T),
    Err(VmError),
}

A result produced by the virtual machine.

Variants

Ok(T)

A produced value.

Err(VmError)

Multiple errors with locations included.

Implementations

impl<T> VmResult<T>

pub fn panic<D>(message: D) -> Self
where D: 'static + BoxedPanic,

Construct a result containing a panic.

pub fn expected<E>(actual: TypeInfo) -> Self
where E: ?Sized + TypeOf,

Construct an expectation error. The actual type received is actual, but we expected E.

pub fn is_ok(&self) -> bool

Test if the result is an ok.

pub fn is_err(&self) -> bool

Test if the result is an error.

pub fn expect(self, msg: &str) -> T

Expect a value or panic.

pub fn unwrap(self) -> T

Unwrap the interior value.

pub fn into_result(self) -> Result<T, VmError>

Convert a VmResult into a Result.

pub fn err<E>(error: E) -> Self
where VmError: From<E>,

Construct a new error from a type that can be converted into a VmError.