Struct rune::runtime::AnyObj

#[repr(C)]
pub struct AnyObj { /* private fields */ }

Our own private dynamic Any implementation.

In contrast to Box<dyn std::any::Any>, this allows for storing a raw pointer directly in the object to avoid one level of indirection. Otherwise it’s equivalent.

Implementations

impl AnyObj

pub fn new<T>(data: T) -> Result<Self>
where T: Any,

Construct a new any from the original any.

pub unsafe fn from_ref<T>(data: &T) -> Self
where T: Any,

Construct an Any that wraps a pointer.

Safety

Caller must ensure that the returned AnyObj doesn’t outlive the reference it is wrapping.

This would be an example of incorrect use:

use rune::Any;
use rune::runtime::AnyObj;

#[derive(Any)]
struct Foo(u32);

let mut v = Foo(1u32);
let any = unsafe { AnyObj::from_ref(&v) };

drop(v);

// any use of `any` beyond here is undefined behavior.

Examples

use rune::Any;
use rune::runtime::AnyObj;

#[derive(Any)]
struct Foo(u32);

let mut v = Foo(1u32);

let any = unsafe { AnyObj::from_ref(&mut v) };
let b = any.downcast_borrow_ref::<Foo>().unwrap();
assert_eq!(b.0, 1u32);

pub unsafe fn from_deref<T>(data: T) -> Result<Self>
where T: Deref, T::Target: Any,

Construct an Any that wraps a Deref type, behaving as the Target of the Deref implementation

Safety

Caller must ensure that the returned AnyObj doesn’t outlive the dereference target.

Examples

use rune::Any;
use rune::runtime::AnyObj;
use std::cell::RefCell;

#[derive(Any)]
struct Foo(u32);

let mut v = RefCell::new(Foo(1u32));
let mut guard = v.borrow();

let any = unsafe { AnyObj::from_deref(guard)? };

let b = any.downcast_borrow_ref::<Foo>().unwrap();
assert_eq!(b.0, 1u32);

pub unsafe fn from_mut<T>(data: &mut T) -> Self
where T: Any,

Construct an Any that wraps a mutable pointer.

Safety

Caller must ensure that the returned AnyObj doesn’t outlive the reference it is wrapping.

This would be an example of incorrect use:

use rune::Any;
use rune::runtime::AnyObj;

#[derive(Any)]
struct Foo(u32);

let mut v = Foo(1u32);
let any = unsafe { AnyObj::from_mut(&mut v) };

drop(v);

// any use of `any` beyond here is undefined behavior.

Examples

use rune::Any;
use rune::runtime::{AnyObj, VmResult};

#[derive(Any)]
struct Foo(u32);

let mut v = Foo(1u32);

{
    let mut any = unsafe { AnyObj::from_mut(&mut v) };

    let Ok(v) = any.downcast_borrow_mut::<Foo>() else {
        panic!("Conversion failed");
    };

    v.0 += 1;
}

assert_eq!(v.0, 2);

pub unsafe fn from_deref_mut<T>(data: T) -> Result<Self>
where T: DerefMut, T::Target: Any,

Construct an Any that wraps a DerefMut type, behaving as the Target of the DerefMut implementation

Safety

Caller must ensure that the returned AnyObj doesn’t outlive the dereference target.

Examples

use rune::Any;
use rune::runtime::AnyObj;
use std::cell::RefCell;

#[derive(Any)]
struct Foo(u32);

let mut v = RefCell::new(Foo(1u32));
let mut guard = v.borrow_mut();

let any = unsafe { AnyObj::from_deref_mut(guard)? };

let b = any.downcast_borrow_ref::<Foo>().unwrap();
assert_eq!(b.0, 1u32);

pub unsafe fn new_raw(vtable: &'static AnyObjVtable, data: *const ()) -> Self

Construct a new any with the specified raw components.

Safety

The caller must ensure that the vtable matches up with the data pointer provided. This is primarily public for use in a C ffi.

pub fn is<T>(&self) -> bool
where T: Any,

Returns true if the boxed type is the same as T.

Examples

use rune::Any;
use rune::runtime::AnyObj;

#[derive(Debug, Any)]
struct Foo;

#[derive(Debug, Any)]
struct Other;

let any = AnyObj::new(Foo)?;

assert!(any.is::<Foo>());
assert!(!any.is::<Other>());

pub fn downcast<T>(self) -> Result<T, (AnyObjError, Self)>
where T: Any,

Returns stored value if it is of type T, or None if it isn’t.

Examples

use rune::Any;
use rune::runtime::AnyObj;

#[derive(Debug, PartialEq, Eq, Any)]
struct Thing(u32);

let any = AnyObj::new(Thing(1u32))?;

let Ok(thing) = any.downcast::<Thing>() else {
    panic!("Conversion failed");
};

assert_eq!(thing, Thing(1));

pub fn downcast_borrow_ref<T>(&self) -> Result<&T, AnyObjError>
where T: Any,

Returns some reference to the boxed value if it is of type T, or None if it isn’t.

Examples

use rune::Any;
use rune::runtime::{AnyObj, AnyObjError};

#[derive(Debug, PartialEq, Eq, Any)]
struct Thing(u32);

#[derive(Debug, PartialEq, Eq, Any)]
struct Other;

let any = AnyObj::new(Thing(1u32))?;
assert_eq!(Ok(&Thing(1u32)), any.downcast_borrow_ref::<Thing>());
assert_eq!(Err(AnyObjError::Cast), any.downcast_borrow_ref::<Other>());

pub fn downcast_borrow_mut<T>(&mut self) -> Result<&mut T, AnyObjError>
where T: Any,

Returns some mutable reference to the boxed value if it is of type T, or None if it isn’t.

Examples

use rune::Any;
use rune::runtime::AnyObj;

#[derive(Debug, PartialEq, Eq, Any)]
struct Thing(u32);

let mut any = AnyObj::new(Thing(1u32))?;
any.downcast_borrow_mut::<Thing>().unwrap().0 = 2;
assert_eq!(Ok(&Thing(2u32)), any.downcast_borrow_ref::<Thing>());

pub fn debug(&self, f: &mut Formatter<'_>) -> Result

Debug format the current any type.

pub fn type_name(&self) -> RawStr

Access the underlying type name for the data.

pub fn type_hash(&self) -> Hash

Access the underlying type id for the data.

pub fn type_info(&self) -> TypeInfo

Access full type info for type.