rune::runtime

Struct Future

pub struct Future { /* private fields */ }

A type-erased future that can only be unsafely polled in combination with the virtual machine that created it.

Implementations

impl Future

pub fn is_completed(&self) -> bool

Check if future is completed.

This will prevent it from being used in a select expression.