Struct rune::runtime::Generator

pub struct Generator<T>
where
    T: AsRef<Vm> + AsMut<Vm>,
{ /* private fields */ }

The return value of a function producing a generator.

Functions which contain the yield keyword produces generators.

Examples

use std::ops::Generator;

fn generate() {
    yield 1;
    yield 2;
}

let g = generate();
assert!(g is Generator)

Implementations

impl<T> Generator<T>
where T: AsRef<Vm> + AsMut<Vm>,

pub fn next(&mut self) -> VmResult<Option<Value>>

Get the next value produced by this stream.

pub fn resume(&mut self, value: Value) -> VmResult<GeneratorState>

Resume the generator with a value and get the next generator state.

impl Generator<&mut Vm>

pub fn into_owned(self) -> Generator<Vm>

Convert the current generator into one which owns its virtual machine.

impl Generator<Vm>

pub fn rune_iter(self) -> Iterator

Convert into iterator