Struct rune::runtime::Protocol

#[non_exhaustive]
pub struct Protocol {
    pub name: &'static str,
    pub hash: Hash,
    /* private fields */
}

A built in instance function.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: &'static str

The name of the builtin function.

hash: Hash

The hash of the builtin function.

Implementations

impl Protocol

pub const GET: Protocol = _

The function to access a field.

pub const GET_HASH: Hash = _

pub const SET: Protocol = _

The function to set a field.

pub const SET_HASH: Hash = _

pub const INDEX_GET: Protocol = _

The function to access an index.

pub const INDEX_GET_HASH: Hash = _

pub const INDEX_SET: Protocol = _

The function to set an index.

pub const INDEX_SET_HASH: Hash = _

pub const PARTIAL_EQ: Protocol = _

Check two types for partial equality.

pub const PARTIAL_EQ_HASH: Hash = _

pub const EQ: Protocol = _

Check two types for total equality.

pub const EQ_HASH: Hash = _

pub const PARTIAL_CMP: Protocol = _

Perform an partial comparison between two values.

pub const PARTIAL_CMP_HASH: Hash = _

pub const CMP: Protocol = _

Perform an total comparison between two values.

pub const CMP_HASH: Hash = _

pub const ADD: Protocol = _

The function to implement for the addition operation.

pub const ADD_HASH: Hash = _

pub const ADD_ASSIGN: Protocol = _

The function to implement for the addition assign operation.

pub const ADD_ASSIGN_HASH: Hash = _

pub const SUB: Protocol = _

The function to implement for the subtraction operation.

pub const SUB_HASH: Hash = _

pub const SUB_ASSIGN: Protocol = _

The function to implement for the subtraction assign operation.

pub const SUB_ASSIGN_HASH: Hash = _

pub const MUL: Protocol = _

The function to implement for the multiply operation.

pub const MUL_HASH: Hash = _

pub const MUL_ASSIGN: Protocol = _

The function to implement for the multiply assign operation.

pub const MUL_ASSIGN_HASH: Hash = _

pub const DIV: Protocol = _

The function to implement for the division operation.

pub const DIV_HASH: Hash = _

pub const DIV_ASSIGN: Protocol = _

The function to implement for the division assign operation.

pub const DIV_ASSIGN_HASH: Hash = _

pub const REM: Protocol = _

The function to implement for the remainder operation.

pub const REM_HASH: Hash = _

pub const REM_ASSIGN: Protocol = _

The function to implement for the remainder assign operation.

pub const REM_ASSIGN_HASH: Hash = _

pub const BIT_AND: Protocol = _

The function to implement for the bitwise and operation.

pub const BIT_AND_HASH: Hash = _

pub const BIT_AND_ASSIGN: Protocol = _

The function to implement for the bitwise and assign operation.

pub const BIT_AND_ASSIGN_HASH: Hash = _

pub const BIT_XOR: Protocol = _

The function to implement for the bitwise xor operation.

pub const BIT_XOR_HASH: Hash = _

pub const BIT_XOR_ASSIGN: Protocol = _

The function to implement for the bitwise xor assign operation.

pub const BIT_XOR_ASSIGN_HASH: Hash = _

pub const BIT_OR: Protocol = _

The function to implement for the bitwise or operation.

pub const BIT_OR_HASH: Hash = _

pub const BIT_OR_ASSIGN: Protocol = _

The function to implement for the bitwise xor assign operation.

pub const BIT_OR_ASSIGN_HASH: Hash = _

pub const SHL: Protocol = _

The function to implement for the bitwise shift left operation.

pub const SHL_HASH: Hash = _

pub const SHL_ASSIGN: Protocol = _

The function to implement for the bitwise shift left assign operation.

pub const SHL_ASSIGN_HASH: Hash = _

pub const SHR: Protocol = _

The function to implement for the bitwise shift right operation.

pub const SHR_HASH: Hash = _

pub const SHR_ASSIGN: Protocol = _

The function to implement for the bitwise shift right assign operation.

pub const SHR_ASSIGN_HASH: Hash = _

pub const STRING_DISPLAY: Protocol = _

Protocol function used by template strings.

pub const STRING_DISPLAY_HASH: Hash = _

pub const STRING_DEBUG: Protocol = _

Protocol function used by custom debug impls.

pub const STRING_DEBUG_HASH: Hash = _

pub const INTO_ITER: Protocol = _

Function used to convert an argument into an iterator.

pub const INTO_ITER_HASH: Hash = _

pub const NEXT: Protocol = _

The function to call to continue iteration.

pub const NEXT_HASH: Hash = _

pub const NTH: Protocol = _

The function to call to continue iteration at the nth element.

pub const NTH_HASH: Hash = _

pub const NTH_BACK: Protocol = _

The function to call to continue iteration at the nth element form the back.

pub const NTH_BACK_HASH: Hash = _

pub const SIZE_HINT: Protocol = _

Protocol used when getting the size hint of an iterator.

pub const SIZE_HINT_HASH: Hash = _

pub const LEN: Protocol = _

Protocol used when getting the exact length of an iterator.

pub const LEN_HASH: Hash = _

pub const NEXT_BACK: Protocol = _

Protocol used when cloning a value.

pub const NEXT_BACK_HASH: Hash = _

pub const INTO_FUTURE: Protocol = _

Function used to convert an argument into a future.

Signature: fn(Value) -> Future.

pub const INTO_FUTURE_HASH: Hash = _

pub const INTO_TYPE_NAME: Protocol = _

Coerce a value into a type name. This is stored as a constant.

pub const INTO_TYPE_NAME_HASH: Hash = _

pub const IS_VARIANT: Protocol = _

Function used to test if a value is a specific variant.

Signature: fn(self, usize) -> bool.

pub const IS_VARIANT_HASH: Hash = _

pub const TRY: Protocol = _

Function used for the question mark operation.

Signature: fn(self) -> Result.

Note that it uses the Result like Try uses ControlFlow i.e., for Result::<T, E> it should return Result<T, Result<(), E>>

pub const TRY_HASH: Hash = _

pub const HASH: Protocol = _

Protocol used when calculating a hash.

pub const HASH_HASH: Hash = _

pub const CLONE: Protocol = _

Protocol used when cloning a value.

pub const CLONE_HASH: Hash = _

pub fn from_hash(hash: Hash) -> Option<Protocol>

Look up protocol for the given hash.