rune::runtime

Struct RangeInclusive

pub struct RangeInclusive {
    pub start: Value,
    pub end: Value,
}

Type for an inclusive range expression start..=end.

Examples

let range = 0..=10;

assert!(!range.contains(-10));
assert!(range.contains(5));
assert!(range.contains(10));
assert!(!range.contains(20));

assert!(range is std::ops::RangeInclusive);

Ranges can contain any type:

let range = 'a'..='f';
assert_eq!(range.start, 'a');
range.start = 'b';
assert_eq!(range.start, 'b');
assert_eq!(range.end, 'f');
range.end = 'g';
assert_eq!(range.end, 'g');

Certain ranges can be used as iterators:

let range = 'a'..='e';
assert_eq!(range.iter().collect::<Vec>(), ['a', 'b', 'c', 'd', 'e']);

Rust Examples

use rune::runtime::RangeInclusive;

let start = rune::to_value(1)?;
let end = rune::to_value(10)?;
let _ = RangeInclusive::new(start, end);

Fields
start: Value

The start value of the range.

end: Value

The end value of the range.

Implementations

impl RangeInclusive

pub fn new(start: Value, end: Value) -> Self

Construct a new range.

pub fn iter(&self) -> Result<Value, VmError>

Iterate over the range.

Panics

This panics if the range is not a well-defined range.

Examples

let range = 'a'..='e';
assert_eq!(range.iter().collect::<Vec>(), ['a', 'b', 'c', 'd', 'e']);

Cannot construct an iterator over floats:

let range = 1.0..=2.0;
range.iter()

pub fn into_iter(&self) -> Result<Value, VmError>

Iterate over the range.

Panics

This panics if the range is not a well-defined range.

Examples

let vec = [];

for value in 'a'..='e' {
    vec.push(value);
}

assert_eq!(vec, ['a', 'b', 'c', 'd', 'e']);

Cannot construct an iterator over floats:

for value in 1.0..=2.0 {
}

pub fn partial_eq(&self, other: &Self) -> Result<bool, VmError>

Test the range for partial equality.

Examples

let range = 'a'..='e';
assert!(range == ('a'..='e'));
assert!(range != ('b'..='e'));

let range = 1.0..=2.0;
assert!(range == (1.0..=2.0));
assert!(range != (f64::NAN..=2.0));
assert!((f64::NAN..=2.0) != (f64::NAN..=2.0));

pub fn eq(&self, other: &Self) -> Result<bool, VmError>

Test the range for total equality.

Examples

use std::ops::eq;

let range = 'a'..='e';
assert!(eq(range, 'a'..='e'));
assert!(!eq(range, 'b'..='e'));

pub fn partial_cmp(&self, other: &Self) -> Result<Option<Ordering>, VmError>

Test the range for partial ordering.

Examples

assert!(('a'..='e') < ('b'..='e'));
assert!(('c'..='e') > ('b'..='e'));
assert!(!((f64::NAN..=2.0) > (f64::INFINITY..=2.0)));
assert!(!((f64::NAN..=2.0) < (f64::INFINITY..=2.0)));

pub fn cmp(&self, other: &Self) -> Result<Ordering, VmError>

Test the range for total ordering.

Examples

use std::ops::cmp;
use std::cmp::Ordering;

assert_eq!(cmp('a'..='e', 'b'..='e'), Ordering::Less);
assert_eq!(cmp('c'..='e', 'b'..='e'), Ordering::Greater);