rune::runtime
Struct TupleStruct

pub struct TupleStruct { /* private fields */ }

A tuple with a well-defined type.

Implementations

impl TupleStruct

pub fn rtti(&self) -> &Arc<Rtti>

Access runtime type information.

pub fn data(&self) -> &OwnedTuple

Access underlying data.

pub fn data_mut(&mut self) -> &mut OwnedTuple

Access underlying data mutably.

pub fn type_info(&self) -> TypeInfo

Get type info for the typed tuple.

pub fn get(&self, index: usize) -> Option<&Value>

Get the value at the given index in the tuple.

pub fn get_mut(&mut self, index: usize) -> Option<&mut Value>

Get the mutable value at the given index in the tuple.