rune::runtime
Struct VmError

pub struct VmError { /* private fields */ }

A virtual machine error which includes tracing information.

Implementations

impl VmError

pub fn emit<O>(&self, out: &mut O, sources: &Sources) -> Result<(), EmitError>
where O: WriteColor,

Available on crate feature emit only.

Generate formatted diagnostics capable of referencing source lines and hints.

See prepare for how to use.

impl VmError

pub fn panic<D>(message: D) -> Self
where D: 'static + BoxedPanic,

Construct an error containing a panic.

pub fn expected<E>(actual: TypeInfo) -> Self
where E: ?Sized + TypeOf,

Construct an expectation error. The actual type received is actual, but we expected E.

pub fn at(&self) -> &VmErrorAt

Get the location where the error happened.

pub fn chain(&self) -> &[VmErrorAt]

Get the full backtrace of errors and their corresponding instructions.

pub fn overflow() -> Self

Construct an overflow error.

pub fn first_location(&self) -> Option<&VmErrorLocation>

Get the first error location.