rune::runtime

Trait TypeOf

pub trait TypeOf: TypeHash {
    const STATIC_TYPE_INFO: AnyTypeInfo;
    const PARAMETERS: Hash = Hash::EMPTY;

    // Provided method
    fn type_info() -> TypeInfo { ... }
}

Trait used for Rust types for which we can determine the runtime type of.

Required Associated Constants

const STATIC_TYPE_INFO: AnyTypeInfo

Access diagnostical type information for the current type.

This can be easily converted to a TypeInfo struct which provides human-readable diagnostics that has a reasonable Display and Debug implementation for humans.

See Self::type_info().

Provided Associated Constants

const PARAMETERS: Hash = Hash::EMPTY

Type parameters for the type.

See ParametersBuilder for more information.

Provided Methods

fn type_info() -> TypeInfo

Get type info associated with the current type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.