Trait rune::runtime::UnsafeToMut

pub trait UnsafeToMut {
    type Guard: 'static;

    // Required method
    unsafe fn unsafe_to_mut<'a>(
        value: Value,
    ) -> VmResult<(&'a mut Self, Self::Guard)>;
}

Unsafe to mut coercion.

Required Associated Types

type Guard: 'static

The raw guard returned.

Must only be dropped after the value returned from this function is no longer live.

Required Methods

unsafe fn unsafe_to_mut<'a>( value: Value, ) -> VmResult<(&'a mut Self, Self::Guard)>

Safety

Caller must ensure that the returned reference does not outlive the guard.

Object Safety

This trait is not object safe.