Trait rune::runtime::UnsafeToValue

pub trait UnsafeToValue: Sized {
    type Guard: 'static;

    // Required method
    unsafe fn unsafe_to_value(self) -> VmResult<(Value, Self::Guard)>;
}

Trait for converting types into values.

Required Associated Types

type Guard: 'static

The type used to guard the unsafe value conversion.

Required Methods

unsafe fn unsafe_to_value(self) -> VmResult<(Value, Self::Guard)>

Convert into a value.

Safety

The value returned must not be used after the guard associated with it has been dropped.

Object Safety

This trait is not object safe.