Trait rune::runtime::unit::UnitStorage

pub trait UnitStorage: Sealed + Debug + Default {
    type Iter<'this>: Iterator<Item = (usize, Inst)>
       where Self: 'this;

    // Required methods
    fn end(&self) -> usize;
    fn bytes(&self) -> usize;
    fn iter(&self) -> Self::Iter<'_>;
    fn get(&self, ip: usize) -> Result<Option<(Inst, usize)>, BadInstruction>;
    fn translate(&self, jump: usize) -> Result<usize, BadJump>;
}

Instruction storage used by a Unit.

Required Associated Types

type Iter<'this>: Iterator<Item = (usize, Inst)> where Self: 'this

Iterator over instructions and their corresponding instruction offsets.

Required Methods

fn end(&self) -> usize

Size of unit storage. This can be seen as the instruction pointer which is just beyond the last instruction.

fn bytes(&self) -> usize

Get the number of bytes which is used to store unit bytecode.

fn iter(&self) -> Self::Iter<'_>

Iterate over all instructions.

fn get(&self, ip: usize) -> Result<Option<(Inst, usize)>, BadInstruction>

Get the instruction at the given instruction pointer.

fn translate(&self, jump: usize) -> Result<usize, BadJump>

Translate the given jump offset.

Object Safety

This trait is not object safe.