rune::sync

Struct Arc

pub struct Arc<T, A = Global>
where
    A: Allocator,
    T: ?Sized,
{ /* private fields */ }

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

The type Arc<T> provides shared ownership of a value of type T, allocated in the heap. Invoking clone on Arc produces a new Arc instance, which points to the same allocation on the heap as the source Arc, while increasing a reference count. When the last Arc pointer to a given allocation is destroyed, the value stored in that allocation (often referred to as “inner value”) is also dropped.

Shared references in Rust disallow mutation by default, and Arc is no exception: you cannot generally obtain a mutable reference to something inside an Arc. If you do need to mutate through an Arc, you have several options:

Use interior mutability with synchronization primitives like Mutex, RwLock, or one of the Atomic types.

Use Arc::get_mut when you know your Arc is not shared (has a reference count of 1), which provides direct mutable access to the inner value without any cloning.

use rune::sync::Arc;
use rune::alloc::try_vec;

let mut data = Arc::try_new(try_vec![1, 2, 3])?;

// This will clone the vector only if there are other references to it
Arc::get_mut(&mut data).unwrap().try_push(4)?;

assert_eq!(*data, try_vec![1, 2, 3, 4]);

Note: This type is only available on platforms that support atomic loads and stores of pointers, which includes all platforms that support the std crate but not all those which only support alloc. This may be detected at compile time using #[cfg(target_has_atomic = "ptr")].

Thread Safety

Unlike Rc<T>, Arc<T> uses atomic operations for its reference counting. This means that it is thread-safe. The disadvantage is that atomic operations are more expensive than ordinary memory accesses. If you are not sharing reference-counted allocations between threads, consider using Rc<T> for lower overhead. Rc<T> is a safe default, because the compiler will catch any attempt to send an Rc<T> between threads. However, a library might choose Arc<T> in order to give library consumers more flexibility.

Arc<T> will implement Send and Sync as long as the T implements Send and Sync. Why can’t you put a non-thread-safe type T in an Arc<T> to make it thread-safe? This may be a bit counter-intuitive at first: after all, isn’t the point of Arc<T> thread safety? The key is this: Arc<T> makes it thread safe to have multiple ownership of the same data, but it doesn’t add thread safety to its data. Consider Arc<RefCell<T>>. RefCell<T> isn’t Sync, and if Arc<T> was always Send, Arc<RefCell<T>> would be as well. But then we’d have a problem: RefCell<T> is not thread safe; it keeps track of the borrowing count using non-atomic operations.

In the end, this means that you may need to pair Arc<T> with some sort of std::sync type, usually Mutex<T>.

Breaking cycles with Weak

The downgrade method can be used to create a non-owning Weak pointer. A Weak pointer can be upgraded to an Arc, but this will return None if the value stored in the allocation has already been dropped. In other words, Weak pointers do not keep the value inside the allocation alive; however, they do keep the allocation (the backing store for the value) alive.

A cycle between Arc pointers will never be deallocated. For this reason, Weak is used to break cycles. For example, a tree could have strong Arc pointers from parent nodes to children, and Weak pointers from children back to their parents.

Cloning references

Creating a new reference from an existing reference-counted pointer is done using the Clone trait implemented for Arc<T> and Weak<T>.

use rune::sync::Arc;
use rune::alloc::try_vec;

let foo = Arc::try_new(try_vec![1.0, 2.0, 3.0])?;
// The two syntaxes below are equivalent.
let a = foo.clone();
let b = Arc::clone(&foo);
// a, b, and foo are all Arcs that point to the same memory location

Deref behavior

Arc<T> automatically dereferences to T (via the Deref trait), so you can call T’s methods on a value of type Arc<T>. To avoid name clashes with T’s methods, the methods of Arc<T> itself are associated functions, called using fully qualified syntax:

use rune::sync::Arc;

let my_arc = Arc::try_new(())?;
let my_weak = Arc::downgrade(&my_arc);

Arc<T>’s implementations of traits like Clone may also be called using fully qualified syntax. Some people prefer to use fully qualified syntax, while others prefer using method-call syntax.

use rune::sync::Arc;

let arc = Arc::try_new(())?;
// Method-call syntax
let arc2 = arc.clone();
// Fully qualified syntax
let arc3 = Arc::clone(&arc);

Weak<T> does not auto-dereference to T, because the inner value may have already been dropped.

Examples

Sharing some immutable data between threads:

use std::thread;

use rune::sync::Arc;

let five = Arc::try_new(5)?;

for _ in 0..10 {
    let five = Arc::clone(&five);

    thread::spawn(move || {
        println!("{five:?}");
    });
}

Sharing a mutable AtomicUsize:

use std::sync::atomic::{AtomicUsize, Ordering};
use std::thread;

use rune::sync::Arc;

let val = Arc::try_new(AtomicUsize::new(5))?;

for _ in 0..10 {
    let val = Arc::clone(&val);

    thread::spawn(move || {
        let v = val.fetch_add(1, Ordering::Relaxed);
        println!("{v:?}");
    });
}

Implementations

impl<T> Arc<T>

pub fn new(data: T) -> Arc<T>

👎Deprecated: Use Arc::try_new instead, which uses checked allocations.

Constructs a new Arc<T>.

Panics

Panics if the allocation fails with an Error.

Examples

use rune::sync::Arc;

let five = Arc::new(5);

pub fn try_new(data: T) -> Result<Arc<T>, Error>

Constructs a new Arc<T>.

Examples

use rune::sync::Arc;

let five = Arc::try_new(5)?;

impl<T, A> Arc<T, A>
where A: Allocator,

pub fn try_new_in(data: T, alloc: A) -> Result<Arc<T, A>, Error>

Constructs a new Arc<T> in the provided allocator.

Examples

use rune::sync::Arc;
use rune::alloc::alloc::Global;

let five = Arc::try_new_in(5, Global)?;

impl<T, A> Arc<T, A>
where A: Allocator, T: ?Sized,

pub fn allocator(this: &Arc<T, A>) -> &A

Returns a reference to the underlying allocator.

Note: this is an associated function, which means that you have to call it as Arc::allocator(&a) instead of a.allocator(). This is so that there is no conflict with a method on the inner type.

pub fn into_raw_with_allocator(this: Arc<T, A>) -> (*const T, A)

Consumes the Arc, returning the wrapped pointer and allocator.

To avoid a memory leak the pointer must be converted back to an Arc using Arc::from_raw_in.

Examples

use rune::alloc::prelude::*;
use rune::sync::Arc;
use rune::alloc::alloc::Global;

let x = Arc::try_new_in("hello".try_to_owned()?, Global)?;
let (ptr, alloc) = Arc::into_raw_with_allocator(x);
assert_eq!(unsafe { &*ptr }, "hello");
let x = unsafe { Arc::from_raw_in(ptr, alloc) };
assert_eq!(&*x, "hello");

pub fn as_ptr(this: &Arc<T, A>) -> *const T

Provides a raw pointer to the data.

The counts are not affected in any way and the Arc is not consumed. The pointer is valid for as long as there are strong counts in the Arc.

Examples

use rune::alloc::prelude::*;
use rune::sync::Arc;

let x = Arc::try_new("hello".try_to_owned()?)?;
let y = Arc::clone(&x);
let x_ptr = Arc::as_ptr(&x);
assert_eq!(x_ptr, Arc::as_ptr(&y));
assert_eq!(unsafe { &*x_ptr }, "hello");

pub unsafe fn from_raw_in(ptr: *const T, alloc: A) -> Arc<T, A>

Constructs an Arc<T, A> from a raw pointer.

The raw pointer has the following requirements:

If U is sized, it must have the same size and alignment as T. This is trivially true if U is T.

If U is unsized, its data pointer must have the same size and alignment as T. This is trivially true if Arc<U> was constructed through Arc<T> and then converted to Arc<U> through an unsized coercion.

Note that if U or U’s data pointer is not T but has the same size and alignment, this is basically like transmuting references of different types. See mem::transmute for more information on what restrictions apply in this case.

The raw pointer must point to a block of memory allocated by alloc

The user of from_raw has to make sure a specific value of T is only dropped once.

This function is unsafe because improper use may lead to memory unsafety, even if the returned Arc<T> is never accessed.

Safety

The pointer must point to an instance which has previously ben returned by Arc<T>::into_raw_with_allocator. The allocator that was used must also be compatible.

Examples

use rune::alloc::prelude::*;
use rune::sync::Arc;
use rune::alloc::alloc::Global;

let x = Arc::try_new_in("hello".try_to_owned()?, Global)?;
let (x_ptr, alloc) = Arc::into_raw_with_allocator(x);

unsafe {
    // Convert back to an `Arc` to prevent leak.
    let x = Arc::from_raw_in(x_ptr, alloc);
    assert_eq!(&*x, "hello");

    // Further calls to `Arc::from_raw(x_ptr)` would be memory-unsafe.
}

// The memory was freed when `x` went out of scope above, so `x_ptr` is now dangling!

Convert a slice back into its original array:

use rune::sync::Arc;

let original: &[u8] = &[1, 2, 3];
let x: Arc<[u8]> = Arc::try_from(original)?;
let (x_ptr, alloc) = Arc::into_raw_with_allocator(x);

unsafe {
    let x: Arc<[u8; 3], _> = Arc::from_raw_in(x_ptr.cast::<[u8; 3]>(), alloc);
    assert_eq!(&*x, &[1, 2, 3]);
}

pub fn ptr_eq(this: &Arc<T, A>, other: &Arc<T, A>) -> bool

Returns true if the two Arcs point to the same allocation in a vein similar to ptr::eq. This function ignores the metadata of dyn Trait pointers.

Examples

use rune::sync::Arc;

let five = Arc::try_new(5)?;
let same_five = Arc::clone(&five);
let other_five = Arc::try_new(5)?;

assert!(Arc::ptr_eq(&five, &same_five));
assert!(!Arc::ptr_eq(&five, &other_five));

pub fn get_mut(this: &mut Arc<T, A>) -> Option<&mut T>

Returns a mutable reference into the given Arc, if there are no other Arc or Weak pointers to the same allocation.

Returns None otherwise, because it is not safe to mutate a shared value.

Examples

use rune::sync::Arc;

let mut x = Arc::try_new(3)?;
*Arc::get_mut(&mut x).unwrap() = 4;
assert_eq!(*x, 4);

let _y = Arc::clone(&x);
assert!(Arc::get_mut(&mut x).is_none());

pub unsafe fn get_mut_unchecked(this: &mut Arc<T, A>) -> &mut T

Returns a mutable reference into the given Arc, without any check.

See also get_mut, which is safe and does appropriate checks.

Safety

If any other Arc or Weak pointers to the same allocation exist, then they must not be dereferenced or have active borrows for the duration of the returned borrow, and their inner type must be exactly the same as the inner type of this Rc (including lifetimes). This is trivially the case if no such pointers exist, for example immediately after Arc::new.

Examples

use rune::sync::Arc;
use rune::alloc::String;

let mut x = Arc::try_new(String::new())?;

unsafe {
    Arc::get_mut_unchecked(&mut x).try_push_str("foo")?
}

assert_eq!(*x, "foo");

Other Arc pointers to the same allocation must be to the same type.

use rune::sync::Arc;

let x: Arc<str> = Arc::try_from("Hello, world!")?;
let mut y: Arc<[u8]> = x.clone().try_into()?;

unsafe {
    // this is Undefined Behavior, because x's inner type is str, not [u8]
    Arc::get_mut_unchecked(&mut y).fill(0xff); // 0xff is invalid in UTF-8
}

println!("{}", &*x); // Invalid UTF-8 in a str

Other Arc pointers to the same allocation must be to the exact same type, including lifetimes.

use rune::sync::Arc;

let x: Arc<&str> = Arc::try_new("Hello, world!")?;

{
    let s = String::from("Oh, no!");
    let mut y: Arc<&str> = x.clone();
    unsafe {
        // this is Undefined Behavior, because x's inner type
        // is &'long str, not &'short str
        *Arc::get_mut_unchecked(&mut y) = &s;
    }
}

println!("{}", &*x); // Use-after-free

pub fn is_unique(this: &Arc<T, A>) -> bool

Determine whether this is the unique reference to the underlying data.

Returns true if there are no other Arc or Weak pointers to the same allocation; returns false otherwise.

If this function returns true, then is guaranteed to be safe to call get_mut_unchecked on this Arc, so long as no clones occur in between.

Examples

use rune::sync::Arc;

let x = Arc::try_new(3)?;
assert!(Arc::is_unique(&x));

let y = Arc::clone(&x);
assert!(!Arc::is_unique(&x));
drop(y);

// Weak references also count, because they could be upgraded at any time.
let z = Arc::downgrade(&x);
assert!(!Arc::is_unique(&x));

Pointer invalidation

This function will always return the same value as Arc::get_mut(arc).is_some(). However, unlike that operation it does not produce any mutable references to the underlying data, meaning no pointers to the data inside the Arc are invalidated by the call. Thus, the following code is valid, even though it would be UB if it used Arc::get_mut:

use rune::sync::Arc;

let arc = Arc::try_new(5)?;
let pointer: *const i32 = &*arc;
assert!(Arc::is_unique(&arc));
assert_eq!(unsafe { *pointer }, 5);

Atomic orderings

Concurrent drops to other Arc pointers to the same allocation will synchronize with this call - that is, this call performs an Acquire operation on the underlying strong and weak ref counts. This ensures that calling get_mut_unchecked is safe.

Note that this operation requires locking the weak ref count, so concurrent calls to downgrade may spin-loop for a short period of time.

pub fn downgrade(this: &Arc<T, A>) -> Weak<T, A>
where A: Clone,

Creates a new Weak pointer to this allocation.

Examples

use rune::sync::Arc;

let five = Arc::try_new(5)?;

let weak_five = Arc::downgrade(&five);