rune::alloc
Module iter

Composable external iteration.

Structs

TryCloned

An iterator that clones the elements of an underlying iterator.

Traits

IteratorExt

Iterator extension trait.

TryExtend

Extend a collection with the contents of an iterator.

TryFromIterator

Conversion from an Iterator within the Global allocator.

TryFromIteratorIn

Conversion from an Iterator within a custom allocator A.

TryJoin

Helper trait for joining iterators.