Trait rune::alloc::iter::TryFromIterator

pub trait TryFromIterator<T>: TryFromIteratorIn<T, Global> {
    // Required method
    fn try_from_iter<I>(iter: I) -> Result<Self, Error>
       where I: IntoIterator<Item = T>;
}

Conversion from an Iterator within the Global allocator.

By implementing TryFromIteratorIn for a type, you define how it will be created from an iterator. This is common for types which describe a collection of some kind.

Required Methods

fn try_from_iter<I>(iter: I) -> Result<Self, Error>
where I: IntoIterator<Item = T>,

Creates a value from an iterator within an allocator.

Object Safety

This trait is not object safe.