rune::alloc::iter

Trait TryFromIteratorIn

pub trait TryFromIteratorIn<T, A>: Sized
where
    A: Allocator,
{
    // Required method
    fn try_from_iter_in<I>(iter: I, alloc: A) -> Result<Self, Error>
       where I: IntoIterator<Item = T>;
}

Conversion from an Iterator within a custom allocator A.

By implementing TryFromIteratorIn for a type, you define how it will be created from an iterator. This is common for types which describe a collection of some kind.

Required Methods

fn try_from_iter_in<I>(iter: I, alloc: A) -> Result<Self, Error>
where I: IntoIterator<Item = T>,

Creates a value from an iterator within an allocator.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.