Module rune::alloc::prelude

Prelude for common traits used in combination with this crate which matches the behavior of the std prelude.

Traits

IteratorExt

Iterator extension trait.

OptionExt

Extensions to Option<T>.

TryClone

Fallible TryClone trait.

TryCopy

Marker trait for types which are Copy.

TryExtend

Extend a collection with the contents of an iterator.

TryFromIterator

Conversion from an Iterator within the Global allocator.

TryFromIteratorIn

Conversion from an Iterator within a custom allocator A.

TryToOwned

A generalization of TryClone to borrowed data.

TryToString

A trait for converting a value to a String.

Derive Macros

TryClone