Module rune::alloc::prelude

Prelude for common traits used in combination with this crate which matches the behavior of the std prelude.

Macros

try_format

Creates a String using interpolation of runtime expressions.

try_vec

Structs

Box

A pointer type that uniquely owns a heap allocation of type T.

String

A UTF-8–encoded, growable string.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Traits

IteratorExt

Iterator extension trait.

OptionExt

Extensions to Option<T>.

TryClone

Fallible TryClone trait.

TryCopy

Marker trait for types which are Copy.

TryExtend

Extend a collection with the contents of an iterator.

TryFromIterator

Conversion from an Iterator within the Global allocator.

TryFromIteratorIn

Conversion from an Iterator within a custom allocator A.

TryToOwned

A generalization of TryClone to borrowed data.

TryToString

A trait for converting a value to a String.

Derive Macros

TryClone