Module rune::module

Types used for defining native modules.

A native module is one that provides rune with functions and types through native Rust-based code.

Structs

Async

Marker for async functions.

EnumMut

Access enum metadata mutably.

InternalEnumMut

Access internal enum metadata mutably.

ItemFnMut

Handle to a an item inserted into a module which allows for mutation of item metadata.

ItemMut

Handle to a an item inserted into a module which allows for mutation of item metadata.

Plain

Marker for plain functions.

TypeMut

Handle to a a type inserted into a module which allows for mutation of its metadata.

Variant

Metadata about a variant.

VariantMut

Handle to a a variant inserted into a module which allows for mutation of its metadata.

Traits

Function

Trait used to provide the function function.

FunctionKind

Denotes the kind of a function, allowing the Function trait to be implemented separately for plain and async functions.

InstallWith

Trait to handle the installation of auxilliary functions for a type installed into a module.

InstanceFunction

Trait used to provide the associated_function function.