Module rune::module

Types used for defining native modules.

A native module is one that provides rune with functions and types through native Rust-based code.

Structs

EnumMut

Access enum metadata mutably.

InternalEnumMut

Access internal enum metadata mutably.

ItemFnMut

Handle to a an item inserted into a module which allows for mutation of item metadata.

ItemMut

Handle to a an item inserted into a module which allows for mutation of item metadata.

Module

A Module that is a collection of native functions and types.

ModuleConstantBuilder

Raw function builder as returned by Module::raw_function.

ModuleFunctionBuilder

Function builder as returned by Module::function.

ModuleMetaData

ModuleRawFunctionBuilder

Raw function builder as returned by Module::raw_function.

TraitMut

Handle to a a trait inserted into a module which allows for mutation of its metadata.

TypeMut

Handle to a a type inserted into a module which allows for mutation of its metadata.

VariantMut

Handle to a a variant inserted into a module which allows for mutation of its metadata.

Traits

InstallWith

Trait to handle the installation of auxilliary functions for a type installed into a module.

Type Aliases

ModuleMeta

Type used to collect and store module metadata through the #[rune::module] macro.