Module rune::runtime::format

Types for dealing with formatting specifications.

Structs

AlignmentFromStrError

Error raised when trying to parse an alignment string and it fails.

Flags

Format specification flags.

Format

A format specification, wrapping an inner value.

FormatSpec

A format specification.

TypeFromStrError

Error raised when trying to parse a type string and it fails.

Enums

Alignment

The alignment requested.

Flag

A single flag for format spec.

Type

The type of formatting requested.