Struct rune::runtime::Object

pub struct Object { /* private fields */ }

Struct representing a dynamic anonymous object.

Rust Examples

use rune::alloc::String;

let mut object = rune::runtime::Object::new();
assert!(object.is_empty());

object.insert_value(String::try_from("foo")?, 42).into_result()?;
object.insert_value(String::try_from("bar")?, true).into_result()?;
assert_eq!(2, object.len());

assert_eq!(Some(42), object.get_value("foo").into_result()?);
assert_eq!(Some(true), object.get_value("bar").into_result()?);
assert_eq!(None::<bool>, object.get_value("baz").into_result()?);

Implementations

impl Object

pub fn new() -> Self

Construct a new object.

Examples

let object = Object::new();
object.insert("Hello", "World");

pub fn with_capacity(capacity: usize) -> Result<Self>

Construct a new object with the given capacity.

pub fn len(&self) -> usize

Returns the number of elements in the object.

Examples

let object = Object::with_capacity(16);
object.insert("Hello", "World");
assert_eq!(object.len(), 1);

pub fn is_empty(&self) -> bool

Returns true if the object is empty.

Examples

let object = Object::with_capacity(16);
assert!(object.is_empty());
object.insert("Hello", "World");
assert!(!object.is_empty());

pub fn get<Q>(&self, k: &Q) -> Option<&Value>
where String: Borrow<Q>, Q: Hash + Eq + Ord + ?Sized,

Returns a reference to the value corresponding to the key.

pub fn get_value<Q, T>(&self, k: &Q) -> VmResult<Option<T>>
where String: Borrow<Q>, Q: Hash + Eq + Ord + ?Sized, T: FromValue,

Get the given value at the given index.

pub fn get_mut<Q>(&mut self, k: &Q) -> Option<&mut Value>
where String: Borrow<Q>, Q: Hash + Eq + Ord + ?Sized,

Returns a mutable reference to the value corresponding to the key.

pub fn contains_key<Q>(&self, k: &Q) -> bool
where String: Borrow<Q>, Q: Hash + Eq + Ord + ?Sized,

Returns true if the map contains a value for the specified key.

pub fn remove<Q>(&mut self, k: &Q) -> Option<Value>
where String: Borrow<Q>, Q: Hash + Eq + Ord + ?Sized,

Removes a key from the map, returning the value at the key if the key was previously in the map.

pub fn insert_value<T>(&mut self, k: String, v: T) -> VmResult<()>
where T: ToValue,

Inserts a key-value pair into the dynamic object, converting it as necessary through the ToValue trait.

pub fn insert(&mut self, k: String, v: Value) -> Result<Option<Value>>

Inserts a key-value pair into the map.

If the map did not have this key present, None is returned.

pub fn clear(&mut self)

Clears the object, removing all key-value pairs. Keeps the allocated memory for reuse.

pub fn into_inner(self) -> HashMap<String, Value>

Convert into inner.

pub fn iter(&self) -> Iter<'_, String, Value>

An iterator visiting all key-value pairs in arbitrary order. The iterator element type is (&'a String, &'a Value).

pub fn keys(&self) -> Keys<'_, String, Value>

An iterator visiting all keys in arbitrary order. The iterator element type is &'a String.

pub fn values(&self) -> Values<'_, String, Value>

An iterator visiting all values in arbitrary order. The iterator element type is &'a Value.

pub fn iter_mut(&mut self) -> IterMut<'_, String, Value>

An iterator visiting all key-value pairs in arbitrary order, with mutable references to the values.

The iterator element type is (&'a String, &'a mut Value).

pub fn rune_iter(this: Ref<Self>) -> Iterator

An iterator visiting all keys and values in arbitrary order.

Examples

let object = #{a: 1, b: 2, c: 3};
let vec = [];

for key in object.iter() {
    vec.push(key);
}

vec.sort_by(|a, b| a.0.cmp(b.0));
assert_eq!(vec, [("a", 1), ("b", 2), ("c", 3)]);