Struct rune::runtime::Ref

pub struct Ref<T: ?Sized> { /* private fields */ }

A strong reference to the given type.

Implementations

impl<T: ?Sized> Ref<T>

pub const fn from_static(value: &'static T) -> Ref<T>

Construct a static reference.

pub fn map<U: ?Sized, F>(this: Self, f: F) -> Ref<U>where F: FnOnce(&T) -> &U,

Map the interior reference of an owned mutable value.

Examples

use rune::runtime::{Shared, Ref};

let vec = Shared::<Vec<u32>>::new(vec![1, 2, 3, 4])?;
let vec = vec.into_ref()?;
let value: Ref<[u32]> = Ref::map(vec, |vec| &vec[0..2]);

assert_eq!(&*value, &[1u32, 2u32][..]);

pub fn try_map<U: ?Sized, F>(this: Self, f: F) -> Option<Ref<U>>where F: FnOnce(&T) -> Option<&U>,

Try to map the reference to a projection.

Examples

use rune::runtime::{Shared, Ref};

let vec = Shared::<Vec<u32>>::new(vec![1, 2, 3, 4])?;
let vec = vec.into_ref()?;
let value: Option<Ref<[u32]>> = Ref::try_map(vec, |vec| vec.get(0..2));

assert_eq!(value.as_deref(), Some(&[1u32, 2u32][..]));

pub fn into_raw(this: Self) -> (NonNull<T>, RawRef)

Convert into a raw pointer and associated raw access guard.

Safety

The returned pointer must not outlive the associated guard, since this prevents other uses of the underlying data which is incompatible with the current.