Struct rune::runtime::Ref

pub struct Ref<T: ?Sized> { /* private fields */ }

A strong reference to the given type.

Implementations

impl<T: ?Sized> Ref<T>

pub const fn from_static(value: &'static T) -> Ref<T>

Construct a static reference.

pub fn map<U: ?Sized, F>(this: Self, f: F) -> Ref<U>
where F: FnOnce(&T) -> &U,

Map the interior reference of an owned mutable value.

Examples

use rune::runtime::{Bytes, Ref};
use rune::alloc::try_vec;

let bytes = rune::to_value(Bytes::from_vec(try_vec![1, 2, 3, 4]))?;
let bytes: Ref<Bytes> = rune::from_value(bytes)?;
let value: Ref<[u8]> = Ref::map(bytes, |vec| &vec[0..2]);

assert_eq!(&*value, &[1, 2][..]);

pub fn try_map<U: ?Sized, F>(this: Self, f: F) -> Result<Ref<U>, Ref<T>>
where F: FnOnce(&T) -> Option<&U>,

Try to map the reference to a projection.

Examples

use rune::runtime::{Bytes, Ref};
use rune::alloc::try_vec;

let bytes = rune::to_value(Bytes::from_vec(try_vec![1, 2, 3, 4]))?;
let bytes: Ref<Bytes> = rune::from_value(bytes)?;

let Ok(value) = Ref::try_map(bytes, |bytes| bytes.get(0..2)) else {
    panic!("Conversion failed");
};

assert_eq!(&value[..], &[1, 2][..]);

pub fn into_raw(this: Self) -> (NonNull<T>, RawRef)

Convert into a raw pointer and associated raw access guard.

Safety

The returned pointer must not outlive the associated guard, since this prevents other uses of the underlying data which is incompatible with the current.

pub unsafe fn from_raw(data: NonNull<T>, guard: RawRef) -> Self

Convert a raw reference and guard into a regular reference.

Safety

The caller is responsible for ensuring that the raw reference is associated with the specific pointer.