rune::runtime

Struct Shared

pub struct Shared<T> { /* private fields */ }

A typed wrapper for a reference.

This is identical in layout to AnyObj, but provides a statically type-checked container.

Implementations

impl<T> Shared<T>
where T: Any,

pub fn new(value: T) -> Result<Self>

Construct a new typed shared value.

Examples

use rune::Value;
use rune::runtime::Shared;
use rune::alloc::String;

let string = String::try_from("Hello World")?;
let string = Shared::new(string)?;
let string = Value::from(string);

let string = string.into_any_obj()?;
assert_eq!(string.borrow_ref::<String>()?.as_str(), "Hello World");

pub fn take(self) -> Result<T, AnyObjError>

Take the owned value of type T.

This consumes any live references of the value and accessing them in the future will result in an error.

Errors

This errors if the underlying value is not owned.

pub fn into_ref(self) -> Result<Ref<T>, AnyObjError>

Downcast into an owned value of type Ref<T>.

Errors

This errors in case the underlying value is not owned, non-owned references cannot be coerced into Ref<T>.

pub fn into_mut(self) -> Result<Mut<T>, AnyObjError>

Downcast into an owned value of type Mut<T>.

Errors

This errors in case the underlying value is not owned, non-owned references cannot be coerced into Mut<T>.

pub fn borrow_ref(&self) -> Result<BorrowRef<'_, T>, AnyObjError>

Borrow a shared reference to the value while checking for shared access.

This prevents other exclusive accesses from being performed while the guard returned from this function is live.

pub fn try_borrow_ref(&self) -> Result<Option<BorrowRef<'_, T>>, AccessError>

Try to borrow an shared reference to the value.

Returns None if the value is not T.

This prevents other exclusive accesses from being performed while the guard returned from this function is live.

pub fn borrow_mut(&self) -> Result<BorrowMut<'_, T>, AnyObjError>

Borrow an exclusive reference to the value.

This prevents other accesses from being performed while the guard returned from this function is live.

pub fn try_borrow_mut(&self) -> Result<Option<BorrowMut<'_, T>>, AccessError>

Try to borrow an exlucisve reference to the value.

Returns None if the value is not T.

This prevents other exclusive accesses from being performed while the guard returned from this function is live.

pub fn is_readable(&self) -> bool

Test if the value is sharable.

pub fn is_writable(&self) -> bool

Test if the value is exclusively accessible.

pub fn type_hash(&self) -> Hash

Access the underlying type id for the data.

pub fn type_info(&self) -> TypeInfo

Access full type info for the underlying type.