Struct rune::runtime::Stack

pub struct Stack { /* private fields */ }

The stack of the virtual machine, where all values are stored.

Implementations

impl Stack

pub fn at(&self, addr: InstAddress) -> Result<&Value, StackError>

Access the value at the given frame offset.

Examples

use rune::vm_try;
use rune::Module;
use rune::runtime::{Output, Stack, VmResult, InstAddress};

fn add_one(stack: &mut Stack, addr: InstAddress, args: usize, out: Output) -> VmResult<()> {
    let value = vm_try!(vm_try!(stack.at(addr)).as_integer());
    out.store(stack, value + 1);
    VmResult::Ok(())
}

pub fn at_mut(&mut self, addr: InstAddress) -> Result<&mut Value, StackError>

Get a value mutable at the given index from the stack bottom.

Examples

use rune::vm_try;
use rune::Module;
use rune::runtime::{Output, Stack, VmResult, InstAddress};

fn add_one(stack: &mut Stack, addr: InstAddress, args: usize, out: Output) -> VmResult<()> {
    let mut value = vm_try!(stack.at_mut(addr));
    let number = vm_try!(value.as_integer());
    *value = vm_try!(rune::to_value(number + 1));
    out.store(stack, ());
    VmResult::Ok(())
}

pub fn slice_at( &self, addr: InstAddress, len: usize, ) -> Result<&[Value], SliceError>

Get the slice at the given address with the given length.

Examples

use rune::vm_try;
use rune::Module;
use rune::runtime::{Output, Stack, ToValue, VmResult, InstAddress};

fn sum(stack: &mut Stack, addr: InstAddress, args: usize, out: Output) -> VmResult<()> {
    let mut number = 0;

    for value in vm_try!(stack.slice_at(addr, args)) {
        number += vm_try!(value.as_integer());
    }

    out.store(stack, number);
    VmResult::Ok(())
}

pub fn slice_at_mut( &mut self, addr: InstAddress, len: usize, ) -> Result<&mut [Value], SliceError>

Get the mutable slice at the given address with the given length.

Examples

use rune::vm_try;
use rune::Module;
use rune::runtime::{Output, Stack, VmResult, InstAddress};

fn sum(stack: &mut Stack, addr: InstAddress, args: usize, out: Output) -> VmResult<()> {
    for value in vm_try!(stack.slice_at_mut(addr, args)) {
        let number = vm_try!(value.as_integer());
        *value = vm_try!(rune::to_value(number + 1));
    }

    out.store(stack, ());
    VmResult::Ok(())
}