Trait rune::runtime::GuardedArgs

pub trait GuardedArgs {
    type Guard;

    // Required methods
    unsafe fn unsafe_into_stack(
        self,
        stack: &mut Stack
    ) -> VmResult<Self::Guard>;
    fn count(&self) -> usize;
}

Trait for converting arguments onto the stack.

This can take references, because it is unsafe to call. And should only be implemented in contexts where it can be guaranteed that the references will not outlive the call.

Required Associated Types

type Guard

Guard that when dropped will invalidate any values encoded.

Required Methods

unsafe fn unsafe_into_stack(self, stack: &mut Stack) -> VmResult<Self::Guard>

Encode arguments onto a stack.

Safety

This is implemented for and allows encoding references on the stack. The returned guard must be dropped before any used references are invalidated.

fn count(&self) -> usize

The number of arguments.