Trait rune::runtime::IteratorTrait

pub trait IteratorTrait: 'static {
    // Required methods
    fn size_hint(&self) -> (usize, Option<usize>);
    fn next(&mut self) -> VmResult<Option<Value>>;
}

The trait for interacting with an iterator.

This has a blanket implementation, and is primarily used to restrict the arguments that can be used in Iterator::from.

Required Methods

fn size_hint(&self) -> (usize, Option<usize>)

Size hint of the iterator.

fn next(&mut self) -> VmResult<Option<Value>>

Get the next value out of the iterator.