rune::runtime
Trait UnsafeToRef

pub trait UnsafeToRef {
    type Guard: 'static;

    // Required method
    unsafe fn unsafe_to_ref<'a>(
        value: Value,
    ) -> VmResult<(&'a Self, Self::Guard)>;
}

Unsafe to ref coercion.

Required Associated Types

type Guard: 'static

The raw guard returned.

Must only be dropped after the value returned from this function is no longer live.

Required Methods

unsafe fn unsafe_to_ref<'a>(value: Value) -> VmResult<(&'a Self, Self::Guard)>

Safety

Caller must ensure that the returned reference does not outlive the guard.

Object Safety

This trait is not object safe.