rune::runtime

Trait UnsafeToValue

pub trait UnsafeToValue: Sized {
    type Guard: 'static;

    // Required method
    unsafe fn unsafe_to_value(
        self,
    ) -> Result<(Value, Self::Guard), RuntimeError>;
}

Trait for converting types into values.

Required Associated Types

type Guard: 'static

The type used to guard the unsafe value conversion.

Required Methods

unsafe fn unsafe_to_value(self) -> Result<(Value, Self::Guard), RuntimeError>

Convert into a value.

Safety

The value returned must not be used after the guard associated with it has been dropped.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.