Module rune::runtime::unit

A single execution unit in the rune virtual machine.

A unit consists of a sequence of instructions, and lookaside tables for metadata like function locations.

Re-exports

pub use self::storage::EncodeError;

Structs

ArrayUnit

Unit stored as array of instructions.

BadInstruction

Error indicating that a bad instruction was located at the given instruction pointer.

BadJump

Error indicating that a bad instruction was located at the given instruction pointer.

ByteCodeUnit

Unit stored as byte code, which is a more compact representation than ArrayUnit, but takes more time to execute since it needs to be decoded as it’s being executed.

Logic

Instructions from a single source file.

Unit

Instructions and debug info from a single source file.

Traits

UnitEncoder

Builder trait for unit storage.

UnitStorage

Instruction storage used by a Unit.

Type Aliases

DefaultStorage

Default storage implementation to use.