rune

Struct Mut

pub struct Mut<T: ?Sized> { /* private fields */ }

A strong owned mutable reference to the given type that can be safely dereferenced.

Examples

Constructed from a static value:

use rune::Mut;

let value: Mut<[u8]> = Mut::from_static(&mut [][..]);
assert_eq!(&value[..], b"");

Implementations

impl<T: ?Sized> Mut<T>

pub fn from_static(value: &'static mut T) -> Mut<T>

Construct an owned mutable reference from a static value.

Examples

use rune::Mut;

let value: Mut<[u8]> = Mut::from_static(&mut [][..]);
assert_eq!(&value[..], b"");

pub fn map<U: ?Sized, F>(this: Self, f: F) -> Mut<U>
where F: FnOnce(&mut T) -> &mut U,

Map the interior reference of an owned mutable value.

Examples

use rune::Mut;
use rune::runtime::Bytes;
use rune::alloc::try_vec;

let bytes = rune::to_value(Bytes::from_vec(try_vec![1, 2, 3, 4]))?;
let bytes: Mut<Bytes> = rune::from_value(bytes)?;
let value: Mut<[u8]> = Mut::map(bytes, |bytes| &mut bytes[0..2]);

assert_eq!(&*value, &mut [1, 2][..]);

pub fn try_map<U: ?Sized, F>(this: Self, f: F) -> Result<Mut<U>, Mut<T>>
where F: FnOnce(&mut T) -> Option<&mut U>,

Try to map the mutable reference to a projection.

Examples

use rune::Mut;
use rune::runtime::Bytes;
use rune::alloc::try_vec;

let bytes = rune::to_value(Bytes::from_vec(try_vec![1, 2, 3, 4]))?;
let bytes: Mut<Bytes> = rune::from_value(bytes)?;

let Ok(mut value) = Mut::try_map(bytes, |bytes| bytes.get_mut(0..2)) else {
    panic!("Conversion failed");
};

assert_eq!(&mut value[..], &mut [1, 2][..]);

pub fn into_raw(this: Self) -> (NonNull<T>, RawAnyGuard)

Convert into a raw pointer and associated raw access guard.

Safety

The returned pointer must not outlive the associated guard, since this prevents other uses of the underlying data which is incompatible with the current.

pub unsafe fn from_raw(value: NonNull<T>, guard: RawAnyGuard) -> Self

Convert a raw mutable reference and guard into a regular mutable reference.

Safety

The caller is responsible for ensuring that the raw mutable reference is associated with the specific pointer.