Module rune::workspace

Available on crate feature workspace only.

Types for dealing with workspaces of rune code.

Structs

Build

A prepared build.

BuildError

Failed to build workspace.

Diagnostics

Diagnostics emitted about a workspace.

FatalDiagnostic

A fatal diagnostic in a workspace.

FileSourceLoader

A filesystem-based source loader.

Found

A found item in the workspace.

FoundPackage

A found item in the workspace associated with a package.

Manifest

The manifest of a workspace.

Package

A single package.

WorkspaceError

An error raised when interacting with workspaces.

Enums

Diagnostic

A single workspace diagnostic.

EmitErroremit

Errors that can be raised when formatting diagnostics.

FoundKind

The kind of a found entry.

WorkspaceFilter

A workspace filter which in combination with functions such as Manifest::find_bins can be used to selectively find things in the workspace.

Constants

MANIFEST_FILE

The name of the toplevel manifest Rune.toml.

Traits

SourceLoader

A source loader.

Functions

prepare

Prepare a workspace build.